### 2022-07-27 #4598 @Kevin-Mattheus-Moerman
@obertscheiderfhwn can you check/fix those potentially missing/invalid DOIs? :point_up: You can check the references again by calling: `@editorialbot check references`, and you can call `@editorialbot generate pdf` to update the draft.
