+++ paper.bib
@@ -1,39 +1,7 @@
-﻿@article{groessing,
-author = {Grössing, Harald and Stadlmajer, Natalie and Fauster, Ewald and Fleischmann, Martin and Schledjewski, Ralf},
-year = {2015},
-month = {04},
-pages = {},
-title = {Flow front advancement during composite processing: Predictions from numerical filling simulation tools in comparison with real-world experiments},
-volume = {37},
-journal = {Polymer Composites},
-doi = {10.1002/pc.23474}
-}
-
-
-@mastersthesis{BSc_ISSE,
-  author  = "Leo Heber",
-  title   = "Reinforcement learning for Resin Transfer Moulding",
-  school  = "Institute for Software and Systems Engineering / University of Augsburg",
-  year    = "in preparation for 2022"
-}
-
-@misc{lims,
-  title = {Publications About LIMS},
-  howpublished = {https://sites.udel.edu/lims/3-2/publications-about-lims/},
-  month="September",
-  year = "2021",
-}
-@online{rtmworx,
-  title = {RTM-Worx Documentation, Overview, User Manual and recent publications},
-  url = {http://http://www.polyworx.com/doc/},
-  month="September",
-  year = "2021"  
-}
-@online{nastran,
-  title = {NASA STRuctrual ANalysis (NASTRAN)},
-  url = {https://software.nasa.gov/software/LAR-16804-GS},
-  month="September",
-  year = "2021"  
+﻿@inproceedings{advani,
+  title={Flow and rheology in polymer composites manufacturing},
+  author={Suresh G. Advani},
+  year={1994}
 }
 
 @article{myrtm,
@@ -45,189 +13,11 @@ title = {MyRTM: An approach for the simulation of resin transfer moulding (RTM) 
 journal = {ECCOMAS 2012 - European Congress on Computational Methods in Applied Sciences and Engineering, e-Book Full Papers}
 }
 
-@online{myrtm1,
-  title = {myRTM - Software for simulating the RTM process},
-  url = {https://www.myrtm.ch/},
-  month="September",
-  year = "2021"  
-}
-
-
-@article{isoldi,
-author = {Isoldi, Liercio and Oliveira, Cristiano and Rocha, Luiz and Avila Souza, Jeferson and Amico, Sandro},
-year = {2012},
-month = {06},
-pages = {105-111},
-title = {Three-dimensional numerical modeling of RTM and LRTM processes},
-volume = {34},
-journal = {Journal of the Brazilian Society of Mechanical Sciences and Engineering},
-doi = {10.1590/S1678-58782012000200001}
-}
-
-
-@article{chan,
-author = {Chan, Albert W. and Hwang, Sun-Tak},
-title = {Anisotropic in-plane permeability of fabric media},
-journal = {Polymer Engineering \& Science},
-volume = {31},
-number = {16},
-pages = {1233-1239},
-doi = {https://doi.org/10.1002/pen.760311613},
-url = {https://onlinelibrary.wiley.com/doi/abs/10.1002/pen.760311613},
-eprint = {https://onlinelibrary.wiley.com/doi/pdf/10.1002/pen.760311613},
-abstract = {Abstract An elegant procedure is proposed for obtaining components of the orthotropic or anisotropic in-plane permeability tensor from experimental data on flow front position and time. A radial flow geometry allows the shape of the advancing flow front to be dictated by the in-plane permeability of the fabric media. The directional permeabilities in the fabric plane are directly calculated from fluid and fabric properties together with data from the mold filling experiment (resin injection pressure and flow front position with time). The simplicity of the apparatus and proposed analytical procedure permit easy testing and comparison of different types of fibrous media.},
[…]
