+++ paper.bib
@@ -1,8 +1,13 @@
-﻿@inproceedings{advani,
-  title={Flow and rheology in polymer composites manufacturing},
-  author={Suresh G. Advani},
-  year={1994}
+﻿@inbook{advani,
+  author    = "Advani, Suresh G. and Bruschke, Michiel V. and Parnas, Richard S.",
+  title     = "Resin Transfer Molding Flow Phenomena in Polymeric Composites",
+  booktitle = "Flow and rheology in polymer composites manufacturing",
+  year      = "1994",
+  publisher = "Elsevier Science B.V.",
+  address   = "Amsterdam, The Netherlands",
+  pages     = "465--515"
 }
+
 
 @article{myrtm,
 author = {Barandun, Gion and Henne, Markus and Giger, E. and Arbter, R.},
