### 2022-08-24 #4598 @Fei-Tao
@helgee This paper is related to composites manufacturing, which belongs to aerospace engineering in general. if you'd like to review this submission, please comment here! Feel free to unsubscribe from this issue if you are not interested.
