### 2022-08-31 #4598 @Fei-Tao
@ketch  since finite area method is a variant of the finite volume method, I thought you have the expertise on this topic. Thank you for your prompt response. Have a good day.
