### 2022-09-15 #4763 @editorialbot
[REVIEW]: RTMsim - A Julia module for filling simulations in Resin Transfer Moulding with the Finite Area Method
**Submitting author:** @obertscheiderfhwn (<a href="http://orcid.org/0000-0002-0023-497X">Christof Obertscheider</a>)
**Repository:** https://github.com/obertscheiderfhwn/RTMsim
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.1
