### 2022-10-16 b34090d
Required changes due to package changes in Makie and GLMakie
