### 2022-10-16 e234a73
Numerical_gradient method=3 for run time optimization can give error if inlet is only one cell
