### 2022-10-16 ffa7754
Fixed (underdetermined) numerical gradient calculation if only one neighbouring cell is available
