+++ rtmsim.jl
@@ -611,8 +611,8 @@ module rtmsim
             for ind in 1:N
                 if celltype[ind]==1  || celltype[ind]==-3; 
                     #Pressure gradient calculation
-                    dpdx,dpdy=numerical_gradient(1,ind,p_old,cellneighboursarray,cellcentertocellcenterx,cellcentertocellcentery);
-                    #dpdx,dpdy=numerical_gradient(3,ind,p_old,cellneighboursarray,cellcentertocellcenterx,cellcentertocellcentery);
+                    #dpdx,dpdy=numerical_gradient(1,ind,p_old,cellneighboursarray,cellcentertocellcenterx,cellcentertocellcentery);
+                    dpdx,dpdy=numerical_gradient(3,ind,p_old,cellneighboursarray,cellcentertocellcenterx,cellcentertocellcentery);
                     
                     #FV scheme for rho,u,v,vof conservation laws
                     cellneighboursline=cellneighboursarray[ind,:];
@@ -811,9 +811,15 @@ module rtmsim
                 Amat[i_neighbour,2]=cellcentertocellcentery[ind,i_neighbour]
                 bvec[i_neighbour]=p_old[i_A]-p_old[i_P];
             end
-            xvec=Amat[1:len_cellneighboursline,:]\bvec[1:len_cellneighboursline];
-            dpdx=xvec[1];
-            dpdy=xvec[2];
+
+            if len_cellneighboursline>1;
+                xvec=Amat[1:len_cellneighboursline,:]\bvec[1:len_cellneighboursline];
+                dpdx=xvec[1];
+                dpdy=xvec[2];        
+            else
+                dpdx=0;
+                dpdy=0;
+            end
         elseif i_method==2;
             #least square solution to determine gradient with limiter
             cellneighboursline=cellneighboursarray[ind,:];
@@ -831,9 +837,15 @@ module rtmsim
                 Amat[i_neighbour,2]=wi[i_neighbour]*cellcentertocellcentery[ind,i_neighbour]
                 bvec[i_neighbour]=wi[i_neighbour]*(p_old[i_A]-p_old[i_P]);
             end
-            xvec=Amat[1:len_cellneighboursline,:]\bvec[1:len_cellneighboursline];
-            dpdx=xvec[1];
-            dpdy=xvec[2];
+
+            if len_cellneighboursline>1
+                xvec=Amat[1:len_cellneighboursline,:]\bvec[1:len_cellneighboursline];
+                dpdx=xvec[1];
+                dpdy=xvec[2];            
+            else
+                dpdx=0;
+                dpdy=0;
+            end
         elseif i_method==3;
             #least square solution to determine gradient - runtime optimized
             cellneighboursline=cellneighboursarray[ind,:];
@@ -852,16 +864,21 @@ module rtmsim
             #dpdx=xvec[1];
             #dpdy=xvec[2];
 
-            Aplus=transpose(Amat)*Amat;
-            a=Aplus[1,1]
-            b=Aplus[1,2]
-            c=Aplus[2,1]
-            d=Aplus[2,2] 
-            bvec_mod=transpose(Amat)*bvec
-            inv = 1/(a * d - b * c)
-            # 1 / (ad -bc) * [d -b; -c a]
-            dpdx = inv * d * bvec_mod[1] - inv * b * bvec_mod[2]
-            dpdy = -inv * c * bvec_mod[1] + inv * a * bvec_mod[2]
+            if len_cellneighboursline>1
+                Aplus=transpose(Amat)*Amat;
+                a=Aplus[1,1]
+                b=Aplus[1,2]
+                c=Aplus[2,1]
+                d=Aplus[2,2] 
[…]
