### 2022-10-23 9633862
Update README.md
Add how to leave package manager.
