### 2022-11-11 ccaa911
Bug fix
- Cell boundary face area at inlet and outlet is only determined from preform patch since inlet and outlet have  the default main preform thickness
- Filling overview plot starts with contour plot for t=0
