+++ rtmsim.jl
@@ -640,8 +640,8 @@ module rtmsim
                         n_y=cellfacenormaly[i_P,i_neighbour];
                         vars_P=[rho_P,u_P,v_P,gamma_P];
                         vars_A=[rho_A,u_A,v_A,gamma_A];
-                        meshparameters=[n_x,n_y,A];
-                        if i_A>0 && (celltype[i_A]==1 || celltype[i_A]==-3);  #neighbour is inner or wall cell
+                        if i_A>0 && (celltype[i_A]==1 || celltype[i_A]==-3);  #neighbour is inner or wall cell                            
+                            meshparameters=[n_x,n_y,A];
                             F_rho_num_add,F_u_num_add,F_v_num_add,F_gamma_num_add,F_gamma_num1_add=numerical_flux_function(1,vars_P,vars_A,meshparameters);
                             F_rho_num=F_rho_num+F_rho_num_add;
                             F_u_num=F_u_num+F_u_num_add;
@@ -650,6 +650,8 @@ module rtmsim
                             F_gamma_num1=F_gamma_num1+F_gamma_num1_add;  
                         end       
                         if i_A>0 && (celltype[i_A]==-1 || celltype[i_A]==-2);  #neighbour is pressure inlet or outlet
+                            A=A*cellthickness[i_P]/(0.5*(cellthickness[i_P]+cellthickness[i_A]));
+                            meshparameters=[n_x,n_y,A];
                             if celltype[i_A]==-2;  #pressure outlet
                                 n_dot_u=dot([n_x; n_y],[u_P; v_P]);
                             elseif celltype[i_A]==-1;  #pressure inlet
@@ -2451,13 +2453,14 @@ module rtmsim
         resolution_val=600;
         fig = Figure(resolution=(resolution_val, resolution_val))   
         ax1 = Axis3(fig[1, 1]; aspect=(ax,ay,az), perspectiveness=0.5,viewmode = :fitzoom,title=string("Filling factor at t=", string(round(t_div*t)/t_div) ,"s"))
-        p1=poly!(ax1,connect(xyz, Point{3}), connect(1:length(X), TriangleFace); color=C_gamma[:], strokewidth=1, colorrange=(0,1))
-        if N1>0;
-            p2=poly!(ax1,connect(xyz1, Point{3}), connect(1:length(X1), TriangleFace); color=C1_gamma[:], strokewidth=1, colorrange=(0,1),colormap = (:bone))
-        end
+        #p1=poly!(ax1,connect(xyz, Point{3}), connect(1:length(X), TriangleFace); color=C_gamma[:], strokewidth=1, colorrange=(0,1))
+        #if N1>0;
+        #    p2=poly!(ax1,connect(xyz1, Point{3}), connect(1:length(X1), TriangleFace); color=C1_gamma[:], strokewidth=1, colorrange=(0,1),colormap = (:bone))
+        #end
         hidedecorations!(ax1);
         hidespines!(ax1) 
-        sl_t = Slider(fig[2, 1], range = time_vector[1]:  (time_vector[end]-time_vector[1])/n_pics :time_vector[end], startvalue =  time_vector[end] );
+        #sl_t = Slider(fig[2, 1], range = time_vector[1]:  (time_vector[end]-time_vector[1])/n_pics :time_vector[end], startvalue =  time_vector[end] );
+        sl_t = Slider(fig[2, 1], range = time_vector[1]:  (time_vector[end]-time_vector[1])/n_pics :time_vector[end], startvalue =  time_vector[1] );
         point = lift(sl_t.value) do x           
             if x<0.5*(time_vector[end]+time_vector[1]);
[…]
