### 2023-02-01 #4763 @obertscheiderfhwn
> Hi @obertscheiderfhwn, we did not receive any response from bbrelji. It seems bbrelji is not available. Do you have anyone in mind that could review this submission? We need the comments from two reviewers to consider publishing a paper.
>
> @duhd1993 Would you please continue your review when you are available? Thanks for your time.

I have two colleagues in mind. From the first I know that he is on GitHub (https://github.com/s1st). One of his students worked with an initial version of the RTMsim solver and from his work he knows RTM mould filling simulations. The second one is a more application related reviewer (https://www.ivw.uni-kl.de/de/ueber-uns/organisation/lebenslaeufe/lebenslauf-miro-duhovic).
