### 2023-02-02 #4763 @obertscheiderfhwn
@Fei-Tao, unfortuately I was not able to find somebody from the RTM filling simulations community on GitHub. The closest was the first suggestion.
Alternatively, one could search on GitHub more generally, for example "laminates" or "classical laminate theory". One finding was https://github.com/AndiMb, the developer of https://tu-dresden.de/ing/maschinenwesen/ilr/lft/elamx2/elamx
