### 2023-02-07 #4763 @Fei-Tao
Hi @mrbuche, thanks for your interest. JOSS does not enforce a hard deadline for reviewing.  But it would be great if you could complete the review in a month.
