### 2023-02-07 #4763 @Zitzeronion
Hi @obertscheiderfhwn and @Fei-Tao 👋

Thanks for considering me as a reviewer. It is correct that the project is not exactly within my area of expertise (CFD with OpenFoam and LBM), but I think I can do a review. In the paper OpenFoam is mentioned as one possible solution to the problem I think. Let me know if both of you think I'm qualified @Fei-Tao and @obertscheiderfhwn.
I have to say though that there would be a lot of suggestion 👀 to improve the repository from my side if I should review the project.
