### 2023-02-07 #4763 @mrbuche
I'd be happy to review this @Fei-Tao if you'll have me. I can help a lot with GitHub/CI/etc. stuff too. Maybe we can get this on the Julia registry as well, so it's a cinch to install.
