### 2023-02-07 #4763 @Zitzeronion
> @Zitzeronion, thanks for you reply. Previously, I used OpenFOAM for my filling simulations. Now I'm using the Julia code which is under review here. At the end of the day @Fei-Tao has to nominate the reviewers. Although I'm curious how to improve the repository, don't forget that I'm no software engineer, I'm developing software for engineering applications.

I'm happy to help, let me know @Fei-Tao if I should review the project 😃
