### 2023-02-13 #4763 @mrbuche
@obertscheiderfhwn I am working on my review and I have a couple things to bring up sooner than later.

- Would you consider packaging this software in a way that is consistent with @JuliaRegistries?
  - You should be able to register it there after a bit of packaging work, but even if it wasn't officially registered there, users would have a relatively automatic way to install it from your GitHub repository. This falls under the "Installation Instructions" checkpoint. I recently went through this process and can help if you have questions.
- Can you add contribution guidelines?
  - See "Community guidelines", this should be pretty quick! This is a good checklist to complete: https://github.com/obertscheiderfhwn/RTMsim/community
- Can you automate your tests, or add more?
  - I see there are some good verification tests, but these don't seem to be automated (using GitHub Actions, for example). For verification tests like these, maybe you can get GitHub Actions to check whether the verification tests produce the expected results when someone is trying to PR into the main branch. That way you can avoid having a commit that breaks your verifiability. Some unit tests would be good too if you can think of some!
- The README is a bit long, but that might just be a personal preference. Feel free to ignore this one unless one of the other reviewers or the editor happens to agree.
-  I can't find any documentation of the code apart from the README and paper (like API docs). High-level documentation is great, but low-level for users who want to dive deep into, and especially extend, your software is vital. Could you use something like Documenter.jl to do this? This goes along with the packaging question above, meaning a proper Julia packaging plus a properly organized and documented package will go together in some sense.
