### 2023-02-13 e165ccd
Include GUI as function rtmsim.gui()
