+++ src/rtmsim.jl
@@ -22,7 +22,7 @@
 
 
 module rtmsim
-    using Glob, LinearAlgebra, JLD2, GeometryBasics, GLMakie, Makie, Random, FileIO, ProgressMeter
+    using Glob, LinearAlgebra, JLD2, GeometryBasics, GLMakie, Makie, Random, FileIO, ProgressMeter, NativeFileDialog, Gtk.ShortNames, Gtk.GConstants, Gtk.Graphics, Gtk
     GLMakie.activate!()    
     #using MAT  #temporary output in Matlab mat-format
             
@@ -1767,7 +1767,7 @@ module rtmsim
         if i_mode==1;
             fig = Figure(resolution=(600, 600))
             ax1 = Axis3(fig[1, 1]; aspect=(ax,ay,az), perspectiveness=0.5,viewmode = :fitzoom,title="Mesh")
-            poly!(connect(xyz, Point{3}), connect(1:length(X), TriangleFace); color=C[:], strokewidth=1)
+            poly!(connect(xyz, Makie.Point{3}), connect(1:length(X), TriangleFace); color=C[:], strokewidth=1)
             #hidedecorations!(ax1);
             hidespines!(ax1) 
             display(fig)
@@ -1952,21 +1952,21 @@ module rtmsim
         ay=(deltay+eps_delta)/(mindelta+eps_delta);
         az=(deltaz+eps_delta)/(mindelta+eps_delta);
         ax1 = Axis3(fig[1, 1]; aspect=(ax,ay,az), perspectiveness=0.5,viewmode = :fitzoom,title="Set 1")
-        poly!(connect(xyz, Point{3}), connect(1:length(X), TriangleFace); color=C_patch1[:], strokewidth=1)
+        poly!(connect(xyz, Makie.Point{3}), connect(1:length(X), TriangleFace); color=C_patch1[:], strokewidth=1)
         hidedecorations!(ax1);hidespines!(ax1) 
         if n_patch>=2;
             ax2 = Axis3(fig[1, 2]; aspect=(ax,ay,az), perspectiveness=0.5,viewmode = :fitzoom,title="Set 2")
-            poly!(connect(xyz, Point{3}), connect(1:length(X), TriangleFace); color=C_patch2[:], strokewidth=1)
+            poly!(connect(xyz, Makie.Point{3}), connect(1:length(X), TriangleFace); color=C_patch2[:], strokewidth=1)
             hidedecorations!(ax2);hidespines!(ax2) 
         end
         if n_patch>=3;
             ax3 = Axis3(fig[1, 3]; aspect=(ax,ay,az), perspectiveness=0.5,viewmode = :fitzoom,title="Set 3")
-            poly!(connect(xyz, Point{3}), connect(1:length(X), TriangleFace); color=C_patch3[:], strokewidth=1)
+            poly!(connect(xyz, Makie.Point{3}), connect(1:length(X), TriangleFace); color=C_patch3[:], strokewidth=1)
             hidedecorations!(ax3);hidespines!(ax3) 
         end
         if n_patch>=4;
             ax4 = Axis3(fig[1, 4]; aspect=(ax,ay,az), perspectiveness=0.5,viewmode = :fitzoom,title="Set 4")
-            poly!(connect(xyz, Point{3}), connect(1:length(X), TriangleFace); color=C_patch4[:], strokewidth=1)
+            poly!(connect(xyz, Makie.Point{3}), connect(1:length(X), TriangleFace); color=C_patch4[:], strokewidth=1)
             hidedecorations!(ax4);hidespines!(ax4) 
         end
         display(fig)
[…]
