+++ src/rtmsim.jl
@@ -2590,7 +2590,7 @@ module rtmsim
         r4[4] = RadioButton(r4[3],choices[4]);       push!(f4,r4[4])
 
         #define logo
-        im=Gtk.GtkImage("figures\\rtmsim_logo1_h200px_grey.png")
+        im=Gtk.GtkImage("..\\figures\\rtmsim_logo1_h200px_grey.png")
 
         #assembly elements in grid pattern
         g = GtkGrid()    #Cartesian coordinates, g[column,row]
@@ -2731,7 +2731,7 @@ module rtmsim
             Gtk.destroy(win)
         end
         function h_clicked(w)
-            i=GtkImage("figures\\rtmsim_help.png");
+            i=GtkImage("..\\figures\\rtmsim_help.png");
             w=GtkWindow(i,"Help");
             show(i);
         end
