+++ src/rtmsim_GUI.jl
@@ -1,0 +1,2 @@
+using rtmsim
+rtmsim.gui()
