+++ rtmsim_GUI.jl
@@ -1,284 +1,0 @@
-#include RTMsim
-include("rtmsim.jl")
-import .rtmsim
-
-#include packages
-using Gtk
-using GLMakie
-using Makie
-using NativeFileDialog
-using Gtk.ShortNames, Gtk.GConstants, Gtk.Graphics
-
-#one Gtk window
-win = GtkWindow("RTMsim"); 
-
-#define buttons
-sm=GtkButton("Select mesh file");pm=GtkButton("Plot mesh");ps=GtkButton("Plot sets")
-ss=GtkButton("Start simulation");cs=GtkButton("Continue simulation")
-sel=GtkButton("Select inlet port"); si=GtkButton("Start interactive");ci=GtkButton("Continue interactive");
-sr=GtkButton("Select results file");pr=GtkButton("Plot results")
-po=GtkButton("Plot overview")
-pf=GtkButton("Plot filling")
-q=GtkButton("Quit")
-h=GtkButton("Help")
-in1=GtkButton("Select input file")
-in3=GtkButton("Run with input file")
-
-#define input fields
-in2=GtkEntry(); set_gtk_property!(in2,:text,"inputfiles\\input.txt");
-mf=GtkEntry(); set_gtk_property!(mf,:text,"meshfiles\\mesh_permeameter1_foursets.bdf");
-t=GtkEntry(); set_gtk_property!(t,:text,"200")
-rf=GtkEntry(); set_gtk_property!(rf,:text,"results.jld2")
-r=GtkEntry(); set_gtk_property!(r,:text,"0.01")
-p1_0=GtkEntry(); set_gtk_property!(p1_0,:text,"Set 1");GAccessor.editable(GtkEditable(p1_0),false) 
-p2_0=GtkEntry(); set_gtk_property!(p2_0,:text,"Set 2");GAccessor.editable(GtkEditable(p2_0),false) 
-p3_0=GtkEntry(); set_gtk_property!(p3_0,:text,"Set 3");GAccessor.editable(GtkEditable(p3_0),false) 
-p4_0=GtkEntry(); set_gtk_property!(p4_0,:text,"Set 4");GAccessor.editable(GtkEditable(p4_0),false) 
-par_1=GtkEntry(); set_gtk_property!(par_1,:text,"135000")
-par_2=GtkEntry(); set_gtk_property!(par_2,:text,"100000")
-par_3=GtkEntry(); set_gtk_property!(par_3,:text,"0.06")
-p0_1=GtkEntry(); set_gtk_property!(p0_1,:text,"0.003")
-p0_2=GtkEntry(); set_gtk_property!(p0_2,:text,"0.7")
-p0_3=GtkEntry(); set_gtk_property!(p0_3,:text,"3e-10")
-p0_4=GtkEntry(); set_gtk_property!(p0_4,:text,"1")
-p0_5=GtkEntry(); set_gtk_property!(p0_5,:text,"1")
-p0_6=GtkEntry(); set_gtk_property!(p0_6,:text,"0")
-p0_7=GtkEntry(); set_gtk_property!(p0_7,:text,"0")
-p1_1=GtkEntry(); set_gtk_property!(p1_1,:text,"0.003")
-p1_2=GtkEntry(); set_gtk_property!(p1_2,:text,"0.7")
-p1_3=GtkEntry(); set_gtk_property!(p1_3,:text,"3e-10")
-p1_4=GtkEntry(); set_gtk_property!(p1_4,:text,"1")
-p1_5=GtkEntry(); set_gtk_property!(p1_5,:text,"1")
-p1_6=GtkEntry(); set_gtk_property!(p1_6,:text,"0")
[…]
