+++ docs/make.jl
@@ -1,0 +1,7 @@
+push!(LOAD_PATH,"../src/")
+using Documenter, rtmsim
+
+makedocs(sitename="RTMsim documentation",
+         modules=[rtmsim],
+         pages=["Home" => "index.md"]
+        )
