+++ src/rtmsim.jl
@@ -26,9 +26,45 @@ module rtmsim
     GLMakie.activate!()    
     #using MAT  #temporary output in Matlab mat-format
             
-    function start_rtmsim(inputfilename)
-        # reads the text input file and calls the solver with the read parameters
-        
+    """
+        start_rtmsim(inputfilename)
+
+    Reads the text input file and calls the solver with the read parameters. The complete set of input parameters can be accessed in the input file. The following paragraph shows an example for such an input file:
+    ```
+        1    #i_model 
+        meshfiles\\mesh_permeameter1_foursets.bdf    #meshfilename 
+        200    #tmax 
+        1.01325e5 1.225 1.4 0.06    #p_ref rho_ref gamma mu_resin_val 
+        1.35e5 1.0e5    #p_a_val p_init_val 
+        3e-3 0.7 3e-10 1 1 0 0    #t_val porosity_val K_val alpha_val refdir1_val refdir2_val refdir3_val 
+        3e-3 0.7 3e-10 1 1 0 0    #t1_val porosity1_val K1_val alpha1_val refdir11_val refdir21_val refdir31_val 
+        3e-3 0.7 3e-10 1 1 0 0    #t2_val porosity2_val K2_val alpha2_val refdir12_val refdir22_val refdir32_val
+        3e-3 0.7 3e-10 1 1 0 0    #t3_val porosity3_val K3_val alpha3_val refdir13_val refdir23_val refdir33_val
+        3e-3 0.7 3e-10 1 1 0 0    #t4_val porosity4_val K4_val alpha4_val refdir14_val refdir24_val refdir34_val 
+        1 0 0 0    #patchtype1val patchtype2val patchtype3val patchtype4val 
+        0 results.jld2    #i_restart restartfilename
+        0 0.01    #i_interactive r_p
+        16    #n_pics
+    ```
+
+    Meaning of the variables:
+    - `i_model`: Identifier for physical model (Default value is 1)
+    - `meshfilename`: Mesh filename.
+    - `tmax`: Maximum simulation time.
+    - `p_ref rho_ref gamma mu_resin_val`: Parameters for the equation of state and dynamic viscosity of resin used in the Darcy term.
+    - `p_a_val p_init_val `: Absolut pressure value for injection port and for initial cavity pressure.
+    - `t_val porosity_val K_val alpha_val refdir1_val refdir2_val refdir3_val`: Properties of the cells in the main preform: The vector `(refdir1_val,refdir2_val,refdir3_val)` is projected onto the cell in order to define the first principal cell direction. The second principal cell direction is perpendicular to the first one in the plane spanned by the cell nodes. The principal cell directions are used as the principal permeabilty directions. The cell properties are defined by the thickness `t_val`, the porosity `porosity_val`, the permeability `K_val` in the first principal cell direction, the permeablity `alpha_val` in the second principal direction.
+    - `t1_val porosity1_val K1_val alpha1_val refdir11_val refdir21_val refdir31_val` etc.: Properties for up to four additional cell regions if preform. 
+    - `patchtype1val patchtype2val patchtype3val patchtype4val`: These regions are used to specify the location of the pressure boundary conditions and to specify regions with different permeability, porosity and thickness properties (e.g. for different part thickness and layup or for race tracking which are regions with very high permeability typically at the boundary of the preforms). Vents need not be specified. Parameters `patchtype1val` define the patch type.
[…]
