+++ docs/make.jl
@@ -5,3 +5,7 @@ makedocs(sitename="RTMsim documentation",
          modules=[rtmsim],
          pages=["Home" => "index.md"]
         )
+
+deploydocs(
+    repo = "github.com/obertscheiderfhwn/RTMsim.git",
+)
