+++ docs/make.jl
@@ -1,3 +1,4 @@
+import Pkg; Pkg.add("Documenter")
 push!(LOAD_PATH,"../src/")
 using Documenter, rtmsim
 
