+++ Project.toml
@@ -4,6 +4,7 @@ authors = ["obertscheiderfhwn <christof.obertscheider@fhwn.ac.at>"]
 version = "1.0.3"
 
 [deps]
+Documenter = "e30172f5-a6a5-5a46-863b-614d45cd2de4"
 FileIO = "5789e2e9-d7fb-5bc7-8068-2c6fae9b9549"
 GLMakie = "e9467ef8-e4e7-5192-8a1a-b1aee30e663a"
 GeometryBasics = "5c1252a2-5f33-56bf-86c9-59e7332b4326"
