+++ docs/make.jl
@@ -2,12 +2,6 @@ import Pkg; Pkg.add("Documenter")
 push!(LOAD_PATH,"../src/")
 using Documenter, rtmsim
 
-makedocs(sitename="RTMsim documentation"  
-         #modules=[rtmsim],
-         #pages=["Home" => "index.md"] ,
-         #repo="https://github.com/obertscheiderfhwn/RTMsim/blob/{commit}{path}#{line}"
-        )
+makedocs(sitename="RTMsim documentation" 
+         )
 
-#deploydocs(
-#    repo = "github.com/obertscheiderfhwn/RTMsim.git",
-#)
