### 2023-02-15 aabfc06
Delete inputfiles directory
