+++ inputfiles/input_case4.txt
@@ -1,15 +1,0 @@
-1    #i_model 
-meshfiles\\mesh_permeameter5.bdf    #meshfilename 
-160    #tmax
-1.01325e5 1.225 1.4 0.071    #p_ref rho_ref gamma mu_resin_val 
-1.91e5 1.0e5    #p_a_val p_init_val 
-3.14e-3 0.604 163e-12 0.3086 1 0 0    #t_val porosity_val K_val alpha_val refdir1_val refdir2_val refdir3_val 
-3.14e-3 0.7 3e-10 1 1 0 0    #t1_val porosity1_val K1_val alpha1_val refdir11_val refdir21_val refdir31_val 
-3.14e-3 0.468 28.6e-12 0.122 1 0 0    #t2_val porosity2_val K2_val alpha2_val refdir12_val refdir22_val refdir32_val 
-3.14e-3 0.7 3e-10 1 1 0 0    #t3_val porosity3_val K3_val alpha3_val refdir13_val refdir23_val refdir33_val 
-3.14e-3 0.7 3e-10 1 1 0 0    #t4_val porosity4_val K4_val alpha4_val refdir14_val refdir24_val refdir34_val 
-1 2 0 0    #patchtype1val patchtype2val patchtype3val patchtype4val 
-0 results.jld2    #i_restart restartfilename 
-0 0.01    #i_interactive r_p 
-16    #n_pics
-
