+++ inputfiles/input_case3_finemesh.txt
@@ -1,15 +1,0 @@
-1    #i_model 
-meshfiles\\mesh_annulusfiller1_meshrefinement.bdf    #meshfilename 
-200    #tmax
-1.01325e5 1.225 1.4 0.06    #p_ref rho_ref gamma mu_resin_val 
-1.35e5 1.0e5    #p_a_val p_init_val 
-3e-3 0.7 3e-10 1 1 0 0    #t_val porosity_val K_val alpha_val refdir1_val refdir2_val refdir3_val 
-3e-3 0.7 3e-10 1 1 0 0    #t1_val porosity1_val K1_val alpha1_val refdir11_val refdir21_val refdir31_val 
-3e-3 0.7 3e-10 1 1 0 0    #t2_val porosity2_val K2_val alpha2_val refdir12_val refdir22_val refdir32_val 
-3e-3 0.7 3e-10 1 1 0 0    #t3_val porosity3_val K3_val alpha3_val refdir13_val refdir23_val refdir33_val 
-3e-3 0.7 3e-10 1 1 0 0    #t4_val porosity4_val K4_val alpha4_val refdir14_val refdir24_val refdir34_val 
-1 0 0 0    #patchtype1val patchtype2val patchtype3val patchtype4val 
-0 results.jld2    #i_restart restartfilename 
-0 0.01    #i_interactive r_p 
-16    #n_pics
-
