### 2023-02-15 e4f3522
Delete test directory
