### 2023-02-15 3ec2bcf
Delete paper.bib
