+++ paper.bib
@@ -1,87 +1,0 @@
-﻿@inbook{advani,
-  author    = "Advani, Suresh G. and Bruschke, Michiel V. and Parnas, Richard S.",
-  title     = "Resin Transfer Molding Flow Phenomena in Polymeric Composites",
-  booktitle = "Flow and rheology in polymer composites manufacturing",
-  year      = "1994",
-  publisher = "Elsevier Science B.V.",
-  address   = "Amsterdam, The Netherlands",
-  pages     = "465--515"
-}
-
-
-@article{myrtm,
-author = {Barandun, Gion and Henne, Markus and Giger, E. and Arbter, R.},
-year = {2012},
-month = {01},
-pages = {8717-8729},
-title = {MyRTM: An approach for the simulation of resin transfer moulding (RTM) processes based on Cellular Automata},
-journal = {ECCOMAS 2012 - European Congress on Computational Methods in Applied Sciences and Engineering, e-Book Full Papers}
-}
-
-@article{barandun,
-author = {Barandun, G. and Henne, M. and Keusch, Th. and Wingrove, Ph. and Thompson, E. and Bottome, K. and Mason, P. and Lee, J. A. and Horejsi, K. and Fleischmann, M. and Obertscheider, Chr. and Eleven, E. and Koller, B. and Gazmend, B. and Salimi, A. and Noisternig, J.},
-year = {2013},
-title = {Development of a next generation composite Annulus Filler for Rolls Royce jet engines},
-journal = {SEICO 13 SAMPE EUROPE 34th International Conference and Forum}
-}
-
-@article{Fauster.2019,
- author = {Fauster, Ewald and Berg, David C. and Abliz, Dilmurat and Gr{\"o}ssing, Harald and Meiners, Dieter and Ziegmann, Gerhard and Schledjewski, Ralf},
- year = {2019},
- title = {{Image processing and data evaluation algorithms for reproducible optical in-plane permeability characterization by radial flow experiments}},
- pages = {45--63},
- volume = {53},
- number = {1},
- journal = {{Journal of Composite Materials}},
- doi = {10.1177/0021998318780209}
-}
-
-@article{groessing,
-author = {Grössing, Harald and Stadlmajer, Natalie and Fauster, Ewald and Fleischmann, Martin and Schledjewski, Ralf},
-year = {2015},
-month = {04},
-pages = {},
-title = {Flow front advancement during composite processing: Predictions from numerical filling simulation tools in comparison with real-world experiments},
-volume = {37},
-journal = {Polymer Composites},
-doi = {10.1002/pc.23474}
-}
-@mastersthesis{BSc_ISSE,
-  author  = "Leo Heber",
-  title   = "Reinforcement learning for Resin Transfer Moulding",
-  school  = "Institute for Software and Systems Engineering / University of Augsburg",
-  year    = "in preparation for 2022"
-}
-
-@online{lims,
-  title = {Publications About LIMS},
-  howpublished = {https://sites.udel.edu/lims/3-2/publications-about-lims/},
-  month="September",
-  year = "2021",
-}
-@online{rtmworx,
-  title = {RTM-Worx Documentation, Overview, User Manual and recent publications},
-  url = {http://http://www.polyworx.com/doc/},
-  month="September",
-  year = "2021"  
-}
-
-@article{tukovic,
-title = {A moving mesh finite volume interface tracking method for surface tension dominated interfacial fluid flow},
-journal = {Computers \& Fluids},
-volume = {55},
-pages = {70-84},
-year = {2012},
-issn = {0045-7930},
-doi = {10.1016/j.compfluid.2011.11.003},
[…]
