### 2023-02-15 5000597
Delete paper.md
