+++ paper.md
@@ -1,106 +1,0 @@
----
-title: 'RTMsim - A Julia module for filling simulations in Resin Transfer Moulding with the Finite Area Method'
-tags:
-  - Computational Fluid Dynamics (CFD)
-  - Shell mesh
-  - Resin Transfer moulding (RTM)
-  - Liquid composite moulding (LCM)
-  - Filling simulation
-  - Julia
-authors:
-  - name: Christof Obertscheider^[Corresponding author]
-    affiliation: 1
-  - name: Ewald Fauster
-    affiliation: 2
-affiliations:
- - name: Aerospace Engineering Department, University of Applied Sciences Wiener Neustadt, Johannes-Gutenberg-Straße 3, 2700, Wiener Neustadt, Austria
-   index: 1
- - name: Processing of Composites Group, Department Polymer Engineering and Science, Montanuniversität Leoben, Otto Glöckl-Straße 2, 8700 Leoben, Austria 
-   index: 2
-date: 7 April 2022
-bibliography: paper.bib
----
-
-# Summary
-RTMsim is a robust, easy-to-use and simple-to-extend software tool, implemented in Julia, for RTM filling simulations. A shell mesh, injection pressure, resin viscosity and the parameters describing the preform are required input. The software was validated with results from literature as well as experiments and compared with results from well-established RTM filling simulation tools.
-
-# Statement of need
-Resin Transfer Moulding (RTM) is a manufacturing process for producing thin-walled fiber reinforced polymer composites where dry fibers are placed inside a mould and resin is injected under pressure into the fibrous preform. The thickness of the part is much smaller than the overall dimensions as depicted in \autoref{fig:rtm_pic1}. During mould design, filling simulations can study different manufacturing concepts (i.e. placement of inlet ports and vents) to guarantee complete filling of the part and avoid air entrapment where flow fronts converge. 
-
-![Schematic of a thin-shell RTM geometry for manufacturing a flat plate with linear (a) and radial (b) flow.\label{fig:rtm_pic1}](figures/rtm_pic1.png)
-
-In the past, numerous models have been implemented in different software packages to perform filling simulations for RTM. The used simulation packages can be divided into three groups: 
-
-- General purpose CFD software packages, such as ANSYS Fluent or OpenFOAM
-
-- Commercially available software packages which are tailored for the simulation of the RTM process, such as PAM-RTM, RTM-Worx or LIMS
-
-- Easy-to-use simulation tools such as myRTM
-
-All packages describe the flow on a macroscopic level. The first group models the flow through the porous cavity using volume-averaged Navier-Stokes equations [@groessing]. The second group makes use of some assumptions on flow physics and boundary conditions and at first solves a Laplace equation for the pressure inside the region which is already filled and in a second step calculates the flow velocity field to propagate the flow front, see e.g. [@rtmworx] or [@lims]. In [@groessing] it has been shown that the first and second group render very similar results. myRTM from the third group is easy-to-use and can predict the filling pattern properly but neither predict the filling time correctly nor consider orthotropic preform permeability, see [@myrtm]. Solving conservation laws for fluid flow as in the first group requires a volume mesh of the cavity and consequently the solution is more time-consuming. The second and third group can be solved on a shell mesh where the thickness of the cavity is a property of the cell (similar to porosity and permeability) and slip boundary conditions at the top and bottom walls of the cavity are assumed. 
-
-Based on the analysis of the existing software tools for RTM filling simulations the following functional requirements for a new software tool were derived: 
-
-- The simulation model shall give correct results for filling pattern and filling time. 
-
-- The simulation tool takes only composite-manufacturing related inputs and the simulation shall be robust and independent of numerics-related input. 
-
-- The simulation tool takes a shell model of the geometry as input and the location-dependent properties are assigned directly on the shell elements. 
-
[…]
