### 2023-02-15 3533e1b
Delete start_rtmsim_GUI.jl
