+++ start_rtmsim_GUI.jl
@@ -1,2 +1,0 @@
-using rtmsim
-rtmsim.gui()
