+++ README.md
@@ -1,33 +1,72 @@
 # RTMsim - A Julia module for filling simulations in Resin Transfer Moulding
 
 
-# Statement of need
+## Mould filling simulations in Resin Transfer Moulding
 Resin Transfer Moulding (RTM) is a manufacturing process for producing thin-walled fiber reinforced polymer composites where dry fibers are placed inside a mould and resin is injected under pressure into the fibrous preform. During mould design, filling simulations can study different manufacturing concepts (i.e. placement of inlet ports and vents) to guarantee complete filling of the part and avoid air entrapment where flow fronts converge. 
 
-In the past, numerous models have been implemented in different software packages to perform filling simulations for RTM. The used simulation packages can be divided into three groups: 
-- General purpose CFD software packages, such as ANSYS Fluent or OpenFOAM
-- Commercially available software packages which are tailored for the simulation of the RTM process, such as PAM-RTM, RTM-Worx or LIMS
-- Easy-to-use simulation tools such as myRTM
+RTMsim is a new software tool for RTM filling simulations. The porous cavity is fully described by a mesh file with triangular cells on the part’s mid-surface and cell set definitions. The latter can be used for specifying the location of the pressure injection ports and regions with different preforms by assigning different thickness, porosity and permeability values. Additional equations (e.g. for modeling the degree-of-cure) can either be added with equations of the same type or modifications of existing equations (e.g. for variable cavity thickness as needed for vacuum assisted resin infusion simulations). Several test cases were used for successfully validating the implemented model.
 
-All packages describe the flow on a macroscopic level. The first group models the flow through the porous cavity using volume-averaged Navier-Stokes equations. The second group makes use of some assumptions and at first solves a Laplace equation for the pressure inside the region which is already filled and in a second step calculates the flow velocity field to propagate the flow front. It has been shown that the first and second group render very similar results. myRTM from the third group is easy-to-use and can predict the filling pattern properly but neither predict the filling time correctly nor consider non-isotropic preform permeability. Solving conservation laws for fluid flow as in the first group requires a volume mesh of the cavity and consequently the solution is more time-consuming. The second and third group can be solved on a shell mesh where the thickness of the cavity is a property of the cell (similar to porosity and permeability) and slip boundary conditions at the top and bottom walls of the cavity are assumed. 
 
-Based on the analysis of existing software tools for RTM filling simulations the following functional requirements for a new software tool were derived:
-- The simulation model shall give correct results for filling pattern and filling time.
-- The simulation tool takes only composite-manufacturing related inputs and the simulation shall be robust independent of numerics-related input.
-- The simulation tool takes a shell model of the geometry as input and the location-dependent properties are assigned directly on the shell elements.
-- New functionalities can be implemented by either adding equations of the same type or modifying existing equations.
+## How to get
 
[…]

[…]
