+++ README.md
@@ -64,212 +64,3 @@ If RTMsim is used for research or development, please use the following entry, s
   year =         {2022}
 }
 ```
-
-
-
-
-
-
-
-# Example usage
-
-## Mesh preparation
-
-The new simulation tool does not include mesh generation. A 3-node triangular shell mesh with the pre-defined regions must be generated with a meshing tool before starting the filling simulation. The authors used Altair HyperWorks but also free software tools such as SALOME-MECA, GMSH or NETGEN can be used.
-
-The shell mesh is created on the part's mid-surface. Mid-surface models are often available in composite manufacturing since computational stress analysis for thin-walled parts is performed on the part’s mid-surface too. 
-
-The prepared shell mesh is imported via a text file where nodes, elements and element sets are described in a format similar to the NASTRAN bulk data format. Every line (except for the set definitions) contains ten fields of eight characters each. The first field contains the name of the item. The following paragraph shows an example for such a mesh file:
-```
-SET 1 = 1,2,3,4,5,6,
-        7,8,9,10,11,12,
-        13,14,15,16
-GRID           1             0.0     0.0     0.0
-GRID           3             0.3     0.3     0.0
-GRID           4            -0.3     0.3     0.0
-[...]
-GRID         331        -2.115-2-.113318     0.0
-GRID         332        -.117148.1562872     0.0
-GRID         333        .2271322.1925105     0.0
-CTRIA3         1       0      15       9      16
-CTRIA3         2       0      16      10      19
-CTRIA3         3       0      19      11      17
-[...]
-CTRIA3       586       0     243     302     332
-CTRIA3       587       0     262     333     259
-CTRIA3       588       0     232     259     333
-```
-Nodes are described by the keyword `GRID`, followed by a grid number, followed by a blank and three fields with the x, y and z coordinates of the node. The triangular cells are defined by the keyword `CTRIA3`, followed by a cell number, followed by a zero, followed by the three node numbers which constitute the cell. Nodes and cells need not be sorted nor starting with one. Cell sets are defined by the keyword `SET` followed by ` N = ` and the cell numbers separated by commas. Not more than 6 cell numbers per line. Then continue with the cell numbers in the line below after 8 blanks at the beginning of the line. Up to four sets can be defined. 
-
-## Run simulation
-
-RTMsim is executed with a well-defined list of parameters specified in an input text file, in the GUI or in the function call `rtmsim.rtmsim_rev1` in the terminal. The following figure shows the GUI with explanation of the parameters. The mesh file and all parameters must be specified in SI units.
-
-<img src="figures/rtmsim_help.png">
-
[…]
