+++ README.md
@@ -1,7 +1,7 @@
 # RTMsim - A Julia module for filling simulations in Resin Transfer Moulding
 
 
-## Mould filling simulations in Resin Transfer Moulding
+## Mould filling simulations
 Resin Transfer Moulding (RTM) is a manufacturing process for producing thin-walled fiber reinforced polymer composites where dry fibers are placed inside a mould and resin is injected under pressure into the fibrous preform. During mould design, filling simulations can study different manufacturing concepts (i.e. placement of inlet ports and vents) to guarantee complete filling of the part and avoid air entrapment where flow fronts converge. 
 
 RTMsim is a new software tool for RTM filling simulations. The porous cavity is fully described by a mesh file with triangular cells on the part’s mid-surface and cell set definitions. The latter can be used for specifying the location of the pressure injection ports and regions with different preforms by assigning different thickness, porosity and permeability values. Additional equations (e.g. for modeling the degree-of-cure) can either be added with equations of the same type or modifications of existing equations (e.g. for variable cavity thickness as needed for vacuum assisted resin infusion simulations). Several test cases were used for successfully validating the implemented model.
@@ -9,24 +9,30 @@ RTMsim is a new software tool for RTM filling simulations. The porous cavity is 
 
 ## How to get
 
-### Requirments and installing Julia
-The RTMsim module was developed with Julia version >= 1.8. Julia is a high level open source programming language and it is as easy to use as python or Matlab. 
+### Requirements and installing Julia
+The RTMsim module was developed and tested with Julia version 1.8. Julia is a high level open source programming language and it is as easy to use as python or Matlab. 
 
-First of all you need a Julia installation.  Download Julia from https://julialang.org/downloads/. Install Julia and add an environment variable such that the Julia terminal can be started from the command line.
+First of all you need a Julia installation.  Download Julia from https://julialang.org/downloads/ and install.
+
+On a Windows operating systems add an environment variable such that the Julia terminal can be started from the command line.
 
 
 ### Installing using the Julia Package manager
 Open a Julia terminal. The only thing you have to do is to add the package to your Julia environment with the following commands:
 - `using Pkg`
 - `Pkg.add(url="https://github.com/obertscheiderfhwn/RTMsim")` or `Pkg.add(url="https://github.com/obertscheiderfhwn/RTMsim",rev="1.0.4")` for specific revision `1.0.4`
-- `Pkg.test("rtmsim")`
+- `Pkg.test("rtmsim")` for an automated test.
+
 Alternatively, one can use the package manager with the following commands:
 - Change to package manager with `]` 
 - `add "https://github.com/obertscheiderfhwn/RTMsim"`
+- `test rtmsim`
 - Return with the `backspace` key
 
 
 ## How to use
+
+### Preparation
 For testing the software create a directory and download mesh- and input-files:
 - Create a working directory
 - Figure out the location of the package with `using rtmsim` and afterwards `pathof(rtmsim)`
@@ -37,21 +43,22 @@ Start the GUI in the Julia terminal:
 - `using rtmsim
 - Start the GUI with `rtmsim.gui()`
 
-If you are working on a Windows operating system you can avoid working in the Julia terminal: 
+If you are working on a Windows operating system you can start RTMsim by double clicking: 
 - Copy `start_rtmsim_gui.bat` and `start_rtmsim_gui.jl` from the package folder to the working directory
 - Double click on `start_rtmsim_gui.bat` in the Explorer
 
-You can start a simulation in the GUI:
[…]
