+++ docs/src/index.md
@@ -66,8 +66,9 @@ RTMsim is executed with a well-defined list of parameters specified in an input 
 
 The following figure shows the GUI with explanation of the parameters.
 ```@raw html
-<img src=../assets/figures/rtmsim_help.png>
+<img src=assets/figures/rtmsim_help.png>
 ```
+
 The buttons in the first line on the LHS are used for mesh inspection, i.e. select a mesh file, plot the mesh with bounding box and plot the defined sets. The buttons in the second line on the LHS are used for starting and continuing a filling simulation. Every time the Start or Continue simulation button is pressed, a filling simulation is started. The simulated flow time `tmax`, the patch types and patch properties must be specified before. Every simulation calculates the flow front propagation during the next `tmax` seconds. If started with the Start simulation button, the cavity is empty initially. If started with the Continue simulation button, the results from the previous simulation are taken as initial condition. With the buttons in the third line one can select inlet ports with specified radius interactively in addition to using the defined sets, and start and continue such a simulation. The buttons in the forth line are used for post-processing, i.e. show filling and pressure distribution of a specified output file (final results are saved in results.jld2), plot filling at four equidistant time instances and filling at different time instances which are selected with a slider bar. The buttons in the line on the RHS are used to start the simulation with the parameters from the selected input file.
 
 The complete set of input parameters can be accessed in the input file. The following paragraph shows an example for such an input file:
