+++ docs/src/parameters.md
@@ -1,0 +1,86 @@
+# Input parameters
+RTMsim is executed with a well-defined list of parameters specified in an input text file or in the GUI. The mesh file and all parameters must be specified in SI units.
+
+## GUI 
+The following figure shows the GUI with explanation of the parameters.
+```@raw html
+<img src=assets/figures/rtmsim_help.png>
+```
+
+The buttons in the first line on the LHS are used for mesh inspection, i.e. select a mesh file, plot the mesh with bounding box and plot the defined sets. The buttons in the second line on the LHS are used for starting and continuing a filling simulation. Every time the Start or Continue simulation button is pressed, a filling simulation is started. The simulated flow time `tmax`, the patch types and patch properties must be specified before. Every simulation calculates the flow front propagation during the next `tmax` seconds. If started with the Start simulation button, the cavity is empty initially. If started with the Continue simulation button, the results from the previous simulation are taken as initial condition. With the buttons in the third line one can select inlet ports with specified radius interactively in addition to using the defined sets, and start and continue such a simulation. The buttons in the forth line are used for post-processing, i.e. show filling and pressure distribution of a specified output file (final results are saved in results.jld2), plot filling at four equidistant time instances and filling at different time instances which are selected with a slider bar. The buttons in the line on the RHS are used to start the simulation with the parameters from the selected input file.
+
+## Input file
+The complete set of input parameters can be accessed in the input file. The following paragraph shows an example for such an input file:
+```
+1    #i_model 
+meshfiles\\mesh_permeameter1_foursets.bdf    #meshfilename 
+200    #tmax 
+1.01325e5 1.225 1.4 0.06    #p_ref rho_ref gamma mu_resin_val 
+1.35e5 1.0e5    #p_a_val p_init_val 
+3e-3 0.7 3e-10 1 1 0 0    #t_val porosity_val K_val alpha_val refdir1_val refdir2_val refdir3_val 
+3e-3 0.7 3e-10 1 1 0 0    #t1_val porosity1_val K1_val alpha1_val refdir11_val refdir21_val refdir31_val 
+3e-3 0.7 3e-10 1 1 0 0    #t2_val porosity2_val K2_val alpha2_val refdir12_val refdir22_val refdir32_val
+3e-3 0.7 3e-10 1 1 0 0    #t3_val porosity3_val K3_val alpha3_val refdir13_val refdir23_val refdir33_val
+3e-3 0.7 3e-10 1 1 0 0    #t4_val porosity4_val K4_val alpha4_val refdir14_val refdir24_val refdir34_val 
+1 0 0 0    #patchtype1val patchtype2val patchtype3val patchtype4val 
+0 results.jld2    #i_restart restartfilename
+0 0.01    #i_interactive r_p
+16    #n_pics
+```
+
+Meaning of the variables:
+- `i_model`: Identifier for physical model (Default value is 1)
+- `meshfilename`: Mesh filename.
+- `tmax`: Maximum simulation time.
+- `p_ref rho_ref gamma mu_resin_val`: Parameters for the equation of state and dynamic viscosity of resin used in the Darcy term.
+- `p_a_val p_init_val `: Absolut pressure value for injection port and for initial cavity pressure.
+- `t_val porosity_val K_val alpha_val refdir1_val refdir2_val refdir3_val`: Properties of the cells in the main preform: The vector `(refdir1_val,refdir2_val,refdir3_val)` is projected onto the cell in order to define the first principal cell direction. The second principal cell direction is perpendicular to the first one in the plane spanned by the cell nodes. The principal cell directions are used as the principal permeabilty directions. The cell properties are defined by the thickness `t_val`, the porosity `porosity_val`, the permeability `K_val` in the first principal cell direction, the permeablity `alpha_val` in the second principal direction.
[…]
