+++ src/rtmsim.jl
@@ -29,7 +29,12 @@ module rtmsim
     """
         start_rtmsim(inputfilename)
 
-    Reads the text input file and calls the solver with the read parameters. The complete set of input parameters can be accessed in the input file. The following paragraph shows an example for such an input file:
+    Arguments:
+        - inputfilename :: String
+
+    Reads the text input file and calls the solver with the read parameters. 
+    
+    The complete set of input parameters can be accessed in the input file. The following paragraph shows an example for such an input file:
     ```
         1    #i_model 
         meshfiles\\mesh_permeameter1_foursets.bdf    #meshfilename 
@@ -184,6 +189,23 @@ module rtmsim
                     t4_val,porosity4_val,K4_val,alpha4_val,refdir14_val,refdir24_val,refdir34_val,
                     patchtype1val,patchtype2val,patchtype3val,patchtype4val,i_restart,restartfilename,i_interactive,r_p,n_pics)
     
+    Arguments:
+    - i_model :: Int
+    - meshfile :: string
+    - tmax :: Float
+    - p_ref, rho_ref, gamma, mu_resin_val :: Float
+    - t_val,porosity_val,K_val,alpha_val,refdir1_val,refdir2_val,refdir3_val :: Float
+    - t1_val,porosity1_val,K1_val,alpha1_val,refdir11_val,refdir21_val,refdir31_val :: Float
+    - t2_val,porosity2_val,K2_val,alpha2_val,refdir12_val,refdir22_val,refdir32_val :: Float
+    - t3_val,porosity3_val,K3_val,alpha3_val,refdir13_val,refdir23_val,refdir33_val :: Float
+    - t4_val,porosity4_val,K4_val,alpha4_val,refdir14_val,refdir24_val,refdir34_val :: Float
+    - patchtype1val,patchtype2val,patchtype3val,patchtype4val :: Int
+    - i_restart :: Int
+    - restartfilename :: string
+    - i_interactive :: Int64
+    - r_p :: Float
+    - n_pics :: Int
+
     RTMsim solver with main steps
     - Simulation initialization
     - Read mesh file and prepare patches  
@@ -850,6 +872,13 @@ module rtmsim
     """
         function numerical_gradient(i_method,ind,p_old,cellneighboursarray,cellcentertocellcenterx,cellcentertocellcentery)
 
+    Arguments:
+    - i_method :: Int
+    - ind :: Int
+    - p_old :: Vector{Float}
+    - cellneighoursarray :: Array{Float,2}
+    - cellcentertocellcenterx, cellcentertocellcentery :: Array{Float,2}
+
     Calculates the pressure gradient from the cell values of the neighbouring cells.
     i_method=1 .. Least square solution to determine gradient
             =2 .. Least square solution to determine gradient with limiter
@@ -946,6 +975,11 @@ module rtmsim
     """
         function numerical_flux_function(i_method,vars_P,vars_A,meshparameters)
 
+    Arguments:
+    - i_method :: Int
+    - vars_P, vars_A :: 4-element Vector{Float}
+    - meshparameters :: 3-element Vector{Float}
+
     Evaluates the numerical flux functions at the cell boundaries.
     i_method==1 .. first order upwinding
     """
@@ -993,6 +1027,12 @@ module rtmsim
 
     """
         function numerical_flux_function_boundary(i_method,vars_P,vars_A,meshparameters,n_dot_u)
+
+    Arguments:
+        - i_method :: Int
+        - vars_P, vars_A :: 4-element Vector{Float}
+        - meshparameters :: 3-element Vector{Float}
+        - n_dot_u :: Float
 
     Evaluates the numerical flux functions at the cell boundaries to pressure inlet or outlet.
     i_method==1 .. first order upwinding
@@ -1071,6 +1111,13 @@ module rtmsim
 
[…]
