+++ docs/src/index.md
@@ -4,6 +4,7 @@ RTMsim is a new software tool for RTM mould filling simulations.
 <img src="assets/figures/overview.png"><br>
 ```
 
+## Introduction 
 Resin Transfer Moulding (RTM) is a manufacturing process for producing thin-walled fiber reinforced polymer composites where dry fibers are placed inside a mould and resin is injected under pressure into the fibrous preform. During mould design, filling simulations can study different manufacturing concepts (i.e. placement of inlet ports and vents) to guarantee complete filling of the part and avoid air entrapment where flow fronts converge. 
 
 In the past, numerous models have been implemented in different software packages to perform filling simulations for RTM. The used simulation packages can be divided into three groups: 
@@ -20,6 +21,47 @@ Based on the analysis of existing software tools for RTM filling simulations the
 - New functionalities can be implemented by either adding equations of the same type or modifying existing equations.
 
 RTMsim is a new software tool for RTM filling simulations which fulfills these requirements: Several test cases were used for successfully validating the implemented model. The porous cavity is fully described by a mesh file with triangular cells on the part’s mid-surface and cell set definitions. The latter can be used for specifying the location of the pressure injection ports and regions with different preforms by assigning different thickness, porosity and permeability values. Additional equations (e.g. for modeling the degree-of-cure) can either be added with equations of the same type or modifications of existing equations (e.g. for variable cavity thickness as needed for vacuum assisted resin infusion simulations). 
+
+
+## Physical model and implementation
+Flow through a porous medium is described by a compressible continuity equation, an adiabatic equation of state and a momentum equation with dynamic viscosity of the resin in the Darcy term. In addition, a volume-of-fluid equation is solved to track the flow front but the calculated fluid fraction is not used to calculate a mean density:
+```@raw html
+<img src="assets/figures/theory_003.png"><br>
+```
+
+The physical quantities in the governing equations are functions of space x and time t:
+```@raw html
+<img src="assets/figures/theory_004.png"><br>
+```
+
+It is assumed that no cross-thickness flow takes place. Consequently, the fluid domain can be meshed with only one cell through the thickness or with a shell mesh on the mid-surface, see (a) and (b). The red inlet and optional outlet ports are also part of the mesh, see (c). 
+```@raw html
+<img src="assets/figures/theory_001.png"><br>
+```
+
+In general, flow through curved cavity with junctions must be solved, see (a). The discretized versions of the continuum, momentum and volume-of-fluid equations are solved on the flattened geometry. The neighbouring cells are rotated about the common edge to lie in a plane, see (b) and (c).
+```@raw html
+<img src="assets/figures/theory_002.png"><br>
+```
+
+Since the flow is described on the mid-surface, the velocity vector has only two non-zero components u and v described in a local coordinate system. The local coordinate system is the result of a projection of a reference vector. The reference vector is chosen such that the local x- and y-directions correspond with the first and second principal permeability directions. 
+
+The four conservation laws for mass density, velocity components u and v in the local cell coordinate system and the fluid fraction describe the flow in the locally flattened geometry, see (a). A coordinate transformation for the velocity components must be performed for every neighbouring cell, see (b).
+```@raw html
+<img src="assets/figures/theory_005.png"><br>
+```
+
[…]
