+++ src/rtmsim.jl
@@ -28,11 +28,11 @@ module rtmsim
             
     """
         start_rtmsim(inputfilename)
+        
+    Reads the text input file and calls the solver with the read parameters. 
 
     Arguments:
-        - inputfilename :: String
-
-    Reads the text input file and calls the solver with the read parameters. 
+    - inputfilename :: String
     
     The complete set of input parameters can be accessed in the input file. The following paragraph shows an example for such an input file:
     ```
@@ -189,24 +189,7 @@ module rtmsim
                     t4_val,porosity4_val,K4_val,alpha4_val,refdir14_val,refdir24_val,refdir34_val,
                     patchtype1val,patchtype2val,patchtype3val,patchtype4val,i_restart,restartfilename,i_interactive,r_p,n_pics)
     
-    Arguments:
-    - i_model :: Int
-    - meshfile :: string
-    - tmax :: Float
-    - p_ref, rho_ref, gamma, mu_resin_val :: Float
-    - t_val,porosity_val,K_val,alpha_val,refdir1_val,refdir2_val,refdir3_val :: Float
-    - t1_val,porosity1_val,K1_val,alpha1_val,refdir11_val,refdir21_val,refdir31_val :: Float
-    - t2_val,porosity2_val,K2_val,alpha2_val,refdir12_val,refdir22_val,refdir32_val :: Float
-    - t3_val,porosity3_val,K3_val,alpha3_val,refdir13_val,refdir23_val,refdir33_val :: Float
-    - t4_val,porosity4_val,K4_val,alpha4_val,refdir14_val,refdir24_val,refdir34_val :: Float
-    - patchtype1val,patchtype2val,patchtype3val,patchtype4val :: Int
-    - i_restart :: Int
-    - restartfilename :: string
-    - i_interactive :: Int64
-    - r_p :: Float
-    - n_pics :: Int
-
-    RTMsim solver with main steps
+    RTMsim solver with the following main steps:
     - Simulation initialization
     - Read mesh file and prepare patches  
     - Find neighbouring cells
@@ -227,6 +210,23 @@ module rtmsim
         - Saving of intermediate data
         - (Opional time marching etc. for i_model=2,3,...)
         - Calculation of adaptive time step 
+
+    Arguments:
+    - i_model :: Int
+    - meshfile :: String
+    - tmax :: Float
+    - p_ref, rho_ref, gamma, mu_resin_val :: Float
+    - t_val,porosity_val,K_val,alpha_val,refdir1_val,refdir2_val,refdir3_val :: Float
+    - t1_val,porosity1_val,K1_val,alpha1_val,refdir11_val,refdir21_val,refdir31_val :: Float
+    - t2_val,porosity2_val,K2_val,alpha2_val,refdir12_val,refdir22_val,refdir32_val :: Float
+    - t3_val,porosity3_val,K3_val,alpha3_val,refdir13_val,refdir23_val,refdir33_val :: Float
+    - t4_val,porosity4_val,K4_val,alpha4_val,refdir14_val,refdir24_val,refdir34_val :: Float
+    - patchtype1val,patchtype2val,patchtype3val,patchtype4val :: Int
+    - i_restart :: Int
+    - restartfilename :: String
+    - i_interactive :: Int64
+    - r_p :: Float
+    - n_pics :: Int
 
     Unit tests:
[…]
