+++ CONTRIBUTING.md
@@ -31,4 +31,4 @@ Contributors can propose changes to the code in the repository by creating a pul
   - The changes provide an appropriate and substantial improvement to the repository.
 
 ## Testing
-Please verify that the code still produced the expected results for the basic V&V cases. The validation and verification cases can be executed in the GUI with the input files input_case1_coarsemesh.txt, input_case1_finemesh.txt, input_case2_coarsemesh.txt, input_case2_finemesh.txt, input_case3_coarsemesh.txt, input_case3_finemesh.txt. The input file names are saved in directory inputfiles.
+Please verify that the code still produced the expected results for the basic [V&V cases](https://obertscheiderfhwn.github.io/RTMsim/build/tutorials/). The validation and verification cases can be executed in the GUI with the input files input_case1_coarsemesh.txt, input_case1_finemesh.txt, input_case2_coarsemesh.txt, input_case2_finemesh.txt, input_case3_coarsemesh.txt, input_case3_finemesh.txt. The input file names are saved in directory inputfiles.
