+++ Project.toml
@@ -1,7 +1,7 @@
 name = "rtmsim"
 uuid = "0823a0f5-0d7f-4277-9a9e-54809b83aaae"
 authors = ["obertscheiderfhwn <christof.obertscheider@fhwn.ac.at>"]
-version = "1.0.3"
+version = "1.0.4"
 
 [deps]
 Documenter = "e30172f5-a6a5-5a46-863b-614d45cd2de4"
