+++ docs/make.jl
@@ -2,6 +2,5 @@ import Pkg; Pkg.add("Documenter")
 push!(LOAD_PATH,"../src/")
 using Documenter, rtmsim
 
-makedocs(sitename="RTMsim documentation"
+makedocs(sitename="RTMsim"
          )
-
