### 2023-02-17 #4763 @obertscheiderfhwn
@mrbuche, here is my update:

- Packaging: I still want to keep the name RTMsim instead of RTMsim.jl because I don't want to scare my potential users which might not even have heard about the programming language Julia. Thats the reason for not officially registrating. Nevertheless, I modified the repository such that RTMsim can be added with the Package manager. Installation instructions are updated.
- Contribution guidelines: Updated.
- Test: I added an automated test for running a simulation and plotting the results. Unit tests for functions are given in the documentation on https://obertscheiderfhwn.github.io/RTMsim/build/functions/. I have not yet implemented a verification test on GitHub actions since the V&V cases take too long. Instead, a added a Testing section in the contribution guidelines.
- README: The old README is still available as README_old, but I created a more concise one and included references to the created documentation on https://obertscheiderfhwn.github.io/RTMsim/build/
- Documentation: I started with a low level documentation on https://obertscheiderfhwn.github.io/RTMsim/build/. There you can find (i) a short introduction with background, model and implementation, (ii) a description of the most important functions, (iii) installation instructions, (iv) a description of the input parameters, and (v) the test cases.

Already a big thank you to both reviewers @mrbuche and @zitzeronion. I had a look at your repositories and copied good ideas :blush:
