+++ docs/src/tutorials.md
@@ -7,10 +7,11 @@ Three different test cases are available, successfully validating the Julia impl
 2. Verification of the software tool for radial flow with tilted orthotropic in-plane permeablity: The simulated tilted elliptical flow front is analysed and the calculated orthotropic permeablity is compared with the simulation input.
 3. Comparison of the simulated flow front position for a complex annulus filler-like part with the simulated flow front position from Ansys Fluent and comparison of the simulated filling pattern with results from a myRTM simulation.
 
-The validation and verification cases can be executed in the GUI. The input files for the validation cases are `input_case1_coarsemesh.txt`, `input_case1_finemesh.txt`, `input_case2_coarsemesh.txt`, `input_case2_finemesh.txt`, `input_case3_coarsemesh.txt`, `input_case3_finemesh.txt`. The input file names are saved in directory `inputfiles` in the package. The filling overview is created with the button `Plot overview` in the GUI. 
+The validation and verification cases can be executed in the GUI. The input files for the validation cases are `input_case1_coarsemesh.txt`, `input_case1_finemesh.txt`, `input_case2_coarsemesh.txt`, `input_case2_finemesh.txt`, `input_case3_coarsemesh.txt`, `input_case3_finemesh.txt`. The input files are saved in directory `inputfiles` in the package folder. The filling overview is created with the button `Plot overview` in the GUI. 
 
 The following passages show and explain the simulation results (after rotating the views with LMB) for the test cases.
 
+### V&V case 1 
 Results for coarse and fine mesh of case 1 (Preform size 0.6 x 0.6 x 0.003 m^3, central injection gate with 0.02 m diameter, 35000 Pa injection pressure, dynamic viscosity 0.06 Pas, porosity 0.7, isotropic permeablity 3e-10 m^2):
 ```@raw html
 <img src="../assets/figures/validation_case1a.png"><br>
@@ -20,6 +21,7 @@ Results for coarse and fine mesh of case 1 (Preform size 0.6 x 0.6 x 0.003 m^3, 
 ```
 According to an analytical estimation, the flow front after 50, 100, 150 and 200 s is a circle with radius 0.114, 0.150, 0.178 and 0.200 m. The values for the fine mesh agree well with the results from the analytical formula. The values for the coarse mesh show an error of <15% which decreases significantly in the course of time but the shape is no smooth circle.
 
+### V&V case 2
 Results for coarse and fine mesh of case 2 (Preform size 0.6 x 0.6 x 0.003 m^3, central injection gate with 0.02 m diameter, 35000 Pa injection pressure, dynamic viscosity 0.06 Pas, porosity 0.70, permeability 3.0e-10 m^2 in first principal direction 30° to the horizontal and permeability 1.5e-10 m^2 in second principal direction):
 ```@raw html
 <img src="../assets/figures/validation_case2a.png"><br>
@@ -29,6 +31,7 @@ Results for coarse and fine mesh of case 2 (Preform size 0.6 x 0.6 x 0.003 m^3, 
 ```
 Analyzing the flow front after 200 s, the orthotropic permeability is described by 2.97e-10 m^2, 1.45e-10 m^2 and angle 35° for the fine mesh and 2.48e-10 m^2, 1.18e-10 m^2 and 35° for the coarse mesh. For the fine mesh (2198 cells for a domain with 600 × 600 mm), this reverse engineering shows very good agreement of the calculated orthotropic permeability  with the values used as simulation input. For the coarse mesh (588 cells) the agreement is still acceptable.
 
+### V&V case 3
[…]
