### 2023-02-17 #4763 @mrbuche
@obertscheiderfhwn

- The packaging is great I think! I was able to install just like the instructions say.
- Found the contributing guidelines and the links all work!
- I am getting this error when trying to run the tests after installing via `Pkg`
  ```
  RTMsim started with the following parameters:
  i_model=1
  meshfilename=..\meshfiles\mesh_permeameter1_foursets.bdf
  ERROR: LoadError: File ..\meshfiles\mesh_permeameter1_foursets.bdf not existing
  ```
  sometimes the paths can be funky with Julia packaging. If you want to use relative paths, you might be able to define a [variable](https://github.com/sandialabs/Polymers/blob/main/src/Polymers.jl#L3) in Julia that is one location, and [write paths](https://github.com/sandialabs/Polymers/blob/main/src/physics/single_chain/fjc/thermodynamics/isometric/mod.jl#L113) relative to that location. I linked an example if that ends up helping.
- README  looks great!
- Docs look great!
