### 2023-02-18 b0e57d2
add variable with path in runtests.jl
