### 2023-02-18 #4763 @obertscheiderfhwn
@mrbuche, unfortunately, I was not able to reproduce the error on two different computers. Nevertheless, I updated the file runtests.jl according to your suggestion (with a variable for the location of the package) and uploaded the new version on the GitHub respository (https://github.com/obertscheiderfhwn/RTMsim/blob/main/test/runtests.jl), but no new release yet created.

No changes in module rtmsim since all paths are relative to the pwd or absolute after interactively selecting a filename in the GUI.

Hope it works now.
