### 2023-02-18 #4763 @mrbuche
It works for me if I change, in the line here

https://github.com/obertscheiderfhwn/RTMsim/blob/main/test/runtests.jl#L11

from

`\\meshfiles\\mesh_`

to

`/meshfiles/mesh_`

Not sure if there are similar instances anywhere else. Test runs and passes for me with this change!
