### 2023-02-19 #4763 @Zitzeronion
@mrbuche this is a os issue. In windows the path is given with `\\` while in linux you have `/`.
I guess the easiest solution would be to ask which os the system is using and do a if else with both paths.
I can suggest a solution later today with a PR 😄
