### 2023-02-19 #4763 @obertscheiderfhwn
@mrbuche and @Zitzeronion, problem seems to be identified. Let's test my PR competence. I am currently working on a solution with https://docs.julialang.org/en/v1/base/file/#Base.Filesystem.joinpath in the [runtests.jl](https://github.com/obertscheiderfhwn/RTMsim/blob/main/test/runtests.jl) and in the unit tests in https://obertscheiderfhwn.github.io/RTMsim/build/functions/. Should not impact the main module since all used files are in top level of the pwd or the test is specified in the GUI.
