### 2023-02-19 dc34a19
test and function docu made independent of OS
