### 2023-02-19 #4763 @mrbuche
@obertscheiderfhwn I'm getting similar problems with the GUI (following instructions in README).

```
File inputfiles\input.txt not existing
```

You could probably set this up in a similar way so that you don't have to change directories to where RTMsim is installed either!

I was trying to verify functionality, especially as shown in the paper, and then after that I'm pretty much done!
