### 2023-02-19 0dfad67
updated paths in gui
