### 2023-02-19 #4763 @obertscheiderfhwn
@mrbuche, I have deleted the default entries right of `Select input file` button and `Select mesh file` button in the GUI (until I will have found a smart way to link to the package folder independently of the OS for the default entries; hopefully tomorrow).
![grafik](https://user-images.githubusercontent.com/102595368/219964879-98dc6ea7-e3cc-4f7d-950d-65c52b5f7060.png)
If you click on these buttons and navigate to the package folder (or at the end of the day to your own mesh- and input-files), a proper Linux or Windows path should be shown. Correct?

This should be fine for the intended use described in https://github.com/obertscheiderfhwn/RTMsim#preparation: Change in the Julia terminal to your working directory and then either copy the mesh- and inputfiles from the package installation to your working directory or copy your own mesh- and inputfiles there. Then open the GUI with `rtmsim.gui()` and click on the buttons `Select input file` or `Select mesh file` to select the files.
