### 2023-02-19 #4763 @mrbuche
I'm still having some trouble with the GUI. When I try to load a mesh file, it crashes and this shows:

```
julia> ┌ Warning: Executing sm_clicked:
└ @ Gtk.GLib ~/.julia/packages/Gtk/oo3cW/src/GLib/signals.jl:106
ERROR: AssertionError: xor(prev, current_task() !== g_stack)
Stacktrace:
  [1] g_siginterruptible(f::typeof(yield), cb::Any)
    @ Gtk.GLib ~/.julia/packages/Gtk/oo3cW/src/GLib/signals.jl:209
  [2] g_yield(data::UInt64)
    @ Gtk.GLib ~/.julia/packages/Gtk/oo3cW/src/GLib/signals.jl:237
  [3] uv_dispatch(src::Ptr{Nothing}, callback::Ptr{Nothing}, data::Int64)
    @ Gtk.GLib ~/.julia/packages/Gtk/oo3cW/src/GLib/signals.jl:333
  [4] OpenDialog
    @ ~/.julia/packages/NativeFileDialog/DjWR5/src/wrapper.jl:3 [inlined]
  [5] pick_file(path::String; filterlist::String)
    @ NativeFileDialog ~/.julia/packages/NativeFileDialog/DjWR5/src/NativeFileDialog.jl:55
  [6] (::rtmsim.var"#sm_clicked#90"{Gtk.GtkEntryLeaf})(w::Gtk.GtkButtonLeaf)
    @ rtmsim ~/.julia/packages/rtmsim/7sWGW/src/rtmsim.jl:2804
  [7] (::Gtk.GLib.var"#90#92"{Ptr{Gtk.GLib.GValue}, UInt32, Ptr{Gtk.GLib.GValue}, Vector{Any}, Bool, rtmsim.var"#sm_clicked#90"{Gtk.GtkEntryLeaf}, Ptr{Any}})()
    @ Gtk.GLib ~/.julia/packages/Gtk/oo3cW/src/GLib/signals.jl:87
  [8] g_siginterruptible(f::Gtk.GLib.var"#90#92"{Ptr{Gtk.GLib.GValue}, UInt32, Ptr{Gtk.GLib.GValue}, Vector{Any}, Bool, rtmsim.var"#sm_clicked#90"{Gtk.GtkEntryLeaf}, Ptr{Any}}, cb::Any)
    @ Gtk.GLib ~/.julia/packages/Gtk/oo3cW/src/GLib/signals.jl:216
  [9] GClosureMarshal(closuref::Ptr{Nothing}, return_value::Ptr{Gtk.GLib.GValue}, n_param_values::UInt32, param_values::Ptr{Gtk.GLib.GValue}, invocation_hint::Ptr{Nothing}, marshal_data::Ptr{Nothing})
    @ Gtk.GLib ~/.julia/packages/Gtk/oo3cW/src/GLib/signals.jl:58
 [10] (::Gtk.var"#253#254")()
    @ Gtk ~/.julia/packages/Gtk/oo3cW/src/events.jl:2
 [11] g_sigatom(f::Any)
    @ Gtk.GLib ~/.julia/packages/Gtk/oo3cW/src/GLib/signals.jl:176
 [12] gtk_main()
    @ Gtk ~/.julia/packages/Gtk/oo3cW/src/events.jl:1

signal (11): Segmentation fault
in expression starting at none:0
g_wakeup_signal at /home/mrbuche/.julia/artifacts/1cfe0ebb804cb8b0d7d1e8f98e5cda94b2b31b3d/lib/libglib-2.0.so.0 (unknown line)
unknown function (ip: 0x7f0a00000001)
unknown function (ip: 0x55c1e08a7a3f)
Allocations: 64782063 (Pool: 64768250; Big: 13813); GC: 30
fish: “julia” terminated by signal SIGSEGV (Address boundary error)
```

I am using a Linux machine. It would be fine if the GUI only worked on certain OS, but that would have to be pointed out, and perhaps some instructions for other OS could be included (how go about using the software to the same level without a GUI).

@Zitzeronion have you succeeded with the GUI yet?
