### 2023-02-19 #4763 @Zitzeronion
@mrbuche I haven't had touched the GUI yet, but I'm curious to find out if my windows machine has the same issues.
That said I had no issue running the test case, but was spending some hours on the paper.
