### 2023-02-19 obertscheiderfhwn/RTMsim#2 @Zitzeronion
paper.md suggestions
Dear @obertscheiderfhwn

First I want to point the huge improvements this package has seen in the last few days, congrats to that 👍.
Concerning the review for [JOSS](https://github.com/openjournals/joss-reviews/issues/4763) I have some minor issues with the paper.
See the attached file, my main criticism concerns missing references and some other small things.
Let me know what you think about it.

[paper.md](https://github.com/obertscheiderfhwn/RTMsim/files/10778251/paper.md)
