### 2023-02-19 obertscheiderfhwn/RTMsim#2 @Zitzeronion
I think it's helpful to have some length scales depict at Figure 1.
![rtm_pic1_suggestion](https://user-images.githubusercontent.com/26249811/219981884-5e4452e6-0cb9-4a11-b0aa-685f26c7e71c.png)
