### 2023-02-20 #4763 @obertscheiderfhwn
@mrbuche: I will test the GUI today on a VM with Ubuntu and try to reproduce and fix the issue. If I will fail, I will update the installation guidelines and tutorials with sections for different OS.

@zitzeronion: I will update the paper according to your suggestions.
