+++ README.md
@@ -30,20 +30,20 @@ Alternatively, one can use the package manager with the following commands:
 - Return with the `backspace` key
 
 
-## How to use
+## How to use - Windows
 
 ### Preparation
 For testing the software create a directory and download mesh- and input-files:
-- Create a working directory
-- Figure out the location of the package with `using rtmsim` and afterwards `pathof(rtmsim)`
-- Copy from the package location the folder with the `meshfiles` and the `inputfiles` into the working directory
+- Create a working directory in the File explorer
+- Figure out the location of the package with `using rtmsim` and afterwards `pathof(rtmsim)` in the Julia terminal
+- Copy from the package location the folder with the `meshfiles` and the `inputfiles` into the working directory in the File explorer
 
 Start the GUI in the Julia terminal:
-- In the Julia terminal change to the working directory, for example with `cd("C:\\obertscheiderfhwn\\rtmsim")` where the separation is a `\\` on Windows or `cd("/home/obertscheiderfhwn/rtmsim")` on Linux operating system
+- Change to the working directory, for example with `cd("C:\\obertscheiderfhwn\\rtmsim")` where the separation is a `\\`
 - `using rtmsim`
 - Start the GUI with `rtmsim.gui()`
 
-If you are working on a Windows operating system you can start RTMsim by double clicking: 
+Start the GUI by double clicking: 
 - Copy `start_rtmsim_gui.bat` and `start_rtmsim_gui.jl` from the package folder to the working directory
 - Double click on `start_rtmsim_gui.bat` in the Explorer
 
@@ -51,11 +51,31 @@ If you are working on a Windows operating system you can start RTMsim by double 
 You can start a simulation in the GUI. 
 <br><img src="figures/rtmsim_help.png"><br>
 
-The buttons in the line on the RHS are used to start the simulation with the parameters from the selected input file. For example, click on `Select input file` and select the input.txt in the inputfiles folder from the package installation and click on `Run with input file`. After the simulation is completed, click on `Plot overview`.
+The buttons in the line on the RHS are used to start the simulation with the parameters from the selected input file. For example, click on `Select input file` and select the input.txt in the inputfiles folder (from the package installation or from the copy in your working directory) and click on `Run with input file`. After the simulation is completed, click on `Plot overview`.
 
[…]
