+++ docs/src/tutorials.md
@@ -7,7 +7,10 @@ Three different test cases are available, successfully validating the Julia impl
 2. Verification of the software tool for radial flow with tilted orthotropic in-plane permeablity: The simulated tilted elliptical flow front is analysed and the calculated orthotropic permeablity is compared with the simulation input.
 3. Comparison of the simulated flow front position for a complex annulus filler-like part with the simulated flow front position from Ansys Fluent and comparison of the simulated filling pattern with results from a myRTM simulation.
 
-The validation and verification cases can be executed in the GUI. The input files for the validation cases are `input_case1_coarsemesh.txt`, `input_case1_finemesh.txt`, `input_case2_coarsemesh.txt`, `input_case2_finemesh.txt`, `input_case3_coarsemesh.txt`, `input_case3_finemesh.txt`. The input files are saved in directory `inputfiles` in the package folder. The filling overview is created with the button `Plot overview` in the GUI. 
+The validation and verification cases can be executed in the GUI with the buttons `Run with input file` and `Plot overview` or from the command line with 
+- `rtmsim.start_rtmsim("inputfiles\\input_case1_coarsemesh.txt")` and `rtmsim.plot_overview(-1,-1)` on Windows and
+- `rtmsim.start_rtmsim("inputfiles/input_case1_coarsemesh_linux.txt")` and `rtmsim.plot_overview(-1,-1)` on Linux.
+The input files for the validation cases are `input_case1_coarsemesh.txt`, `input_case1_finemesh.txt`, `input_case2_coarsemesh.txt`, `input_case2_finemesh.txt`, `input_case3_coarsemesh.txt`, `input_case3_finemesh.txt`. The input files are saved in directory `inputfiles` in the package folder. The appropriate file must be selected in GUI or in the function call.
 
 The following passages show and explain the simulation results (after rotating the views with LMB) for the test cases.
 
@@ -42,12 +45,12 @@ Results for coarse and fine mesh of case 3 (Bounding box of part is 0.4 x 0.15 x
 If the position of inlet and outlet ports is investigated, simulations with a coarse mesh and consequently reduced computational time is sufficient since the flow front progagation is predicted properly. The actual filling time can only be predicted with a fine mesh. A mesh refinement study must be performed. With the coarse mesh it takes approximately 50 s longer to reach the same filling state. If the predicted filling with the fine mesh is considered correct, the error with the coarse mesh is approximately 20%. This is in good agreement with the results from case 1 where domain size, coarse and fine mesh are similar and the permeability and injection values are the same. 
 
 
-## Use cases
+## Use cases - Windows
 
 The following three examples show typical use cases how RTMsim is used by engineers:
 
 ### Use case 1 
-This examples shows how to start a simulation with the GUI. Select the mesh file `meshfiles\\mesh_permeameter1_foursets.bdf` and set all other parameters as shown:
[…]
