+++ src/rtmsim.jl
@@ -2823,11 +2823,11 @@ module rtmsim
         #callback functions
         function sm_clicked(w)
             str = pick_file(pwd(),filterlist="bdf");
-            @async set_gtk_property!(mf,:text,str);
+            set_gtk_property!(mf,:text,str);
         end
         function sr_clicked(w)
             str = pick_file(pwd(),filterlist="jld2");
-            @async set_gtk_property!(rf,:text,str);
+            set_gtk_property!(rf,:text,str);
         end
         function pm_clicked(w)
             str = get_gtk_property(mf,:text,String)
@@ -2946,7 +2946,7 @@ module rtmsim
         #end
         function in1_clicked(w)
             str = pick_file(pwd(),filterlist="txt");
-            @async set_gtk_property!(in2,:text,str);
+            set_gtk_property!(in2,:text,str);
         end
         function in3_clicked(w)
             str = get_gtk_property(in2,:text,String)
