### 2023-02-20 8c10c97
updated linux procedure
