+++ src/rtmsim.jl
@@ -38,7 +38,7 @@ module rtmsim
     - `i_model`: Identifier for physical model (Default value is 1)
     - `meshfilename`: Mesh filename.
     - `tmax`: Maximum simulation time.
-    - `p_ref rho_ref gamma mu_resin_val`: Parameters for the equation of state and dynamic viscosity of resin used in the Darcy term.
+    - `p_ref rho_ref gamma mu_resin_val`: Parameters for the adiabatic equation of state and dynamic viscosity of resin used in the Darcy term.
     - `p_a_val p_init_val `: Absolut pressure value for injection port and for initial cavity pressure.
     - `t_val porosity_val K_val alpha_val refdir1_val refdir2_val refdir3_val`: Properties of the cells in the main preform: The vector `(refdir1_val,refdir2_val,refdir3_val)` is projected onto the cell in order to define the first principal cell direction. The second principal cell direction is perpendicular to the first one in the plane spanned by the cell nodes. The principal cell directions are used as the principal permeabilty directions. The cell properties are defined by the thickness `t_val`, the porosity `porosity_val`, the permeability `K_val` in the first principal cell direction, the permeablity `alpha_val` in the second principal direction.
     - `t1_val porosity1_val K1_val alpha1_val refdir11_val refdir21_val refdir31_val` etc.: Properties for up to four additional cell regions if preform. 
@@ -47,9 +47,9 @@ module rtmsim
         - 1 .. the patch represents an inlet gate, where the specified injection pressure level applies
         - 2 .. the patch specifies a preform region
         - 3 .. the patch represents a vent, where the specified initial pressure level applies
-    - `i_restart restartfilename`: Start with new simulation if `0` or continue previous simulation if `1`.
-    - `i_interactive r_p`: Select the inlet ports graphically if i_interactive equal to `1`.
-    - `n_pics`: Number of intermediate output files. Supposed to be a multiple of `4`.
+    - `i_restart restartfilename`: Start with new simulation if `0` or continue previous simulation if `1` from specified file
+    - `i_interactive r_p`: Select the inlet ports graphically if i_interactive equal to `1` and inlet ports have specified radius
+    - `n_pics`: Number of intermediate output files, supposed to be a multiple of `4`
     Entries are separated by one blank.
 
     Unit test:
@@ -219,6 +219,23 @@ module rtmsim
     - i_interactive :: Int64
     - r_p :: Float
     - n_pics :: Int
+
+    Meaning of the variables:
+    - `i_model`: Identifier for physical model (Default value is 1)
+    - `meshfilename`: Mesh filename.
+    - `tmax`: Maximum simulation time.
+    - `p_ref rho_ref gamma mu_resin_val`: Parameters for the adiabatic equation of state and dynamic viscosity of resin used in the Darcy term.
+    - `p_a_val p_init_val `: Absolut pressure value for injection port and for initial cavity pressure.
+    - `t_val porosity_val K_val alpha_val refdir1_val refdir2_val refdir3_val`: Properties of the cells in the main preform: The vector `(refdir1_val,refdir2_val,refdir3_val)` is projected onto the cell in order to define the first principal cell direction. The second principal cell direction is perpendicular to the first one in the plane spanned by the cell nodes. The principal cell directions are used as the principal permeabilty directions. The cell properties are defined by the thickness `t_val`, the porosity `porosity_val`, the permeability `K_val` in the first principal cell direction, the permeablity `alpha_val` in the second principal direction.
[…]
