+++ README.md
@@ -74,8 +74,9 @@ Run a simulation in the Julia terminal:
 - `using rtmsim`
 - Start a simulation with input file if you copied the inputfiles folder to your working directory: `rtmsim.start_rtmsim("inputfiles/input.txt")` or
 - Start a simulation by calling the solver with all arguments: `rtmsim.rtmsim_rev1(1,"meshfiles/mesh_permeameter1_foursets.bdf",200, 101325,1.225,1.4,0.06, 1.35e5,1.00e5, 3e-3,0.7,3e-10,1,1,0,0, 3e-3,0.7,3e-10,1,1,0,0, 3e-3,0.7,3e-11,1,1,0,0, 3e-3,0.7,3e-11,1,1,0,0, 3e-3,0.7,3e-9,1,1,0,0, 1,2,2,2,0,"results.jld2",0,0.01,16)`
+- Create contour plots of filling at different time instances: `rtmsim.plot_filling(-1,-1)`
 
-Click [here](https://obertscheiderfhwn.github.io/RTMsim/build/parameters/#Input-file) for a documentation of the input file and the meaning of all the parameters  and [here](https://obertscheiderfhwn.github.io/RTMsim/build/tutorials/) for tutorials (with typical use cases). 
+Click [here](https://obertscheiderfhwn.github.io/RTMsim/build/functions/#rtmsim.rtmsim_rev1-NTuple{53,%20Any}) for a description of the parameters, click [here](https://obertscheiderfhwn.github.io/RTMsim/build/parameters/#Input-file) for a description of the input file and [here](https://obertscheiderfhwn.github.io/RTMsim/build/tutorials/) for tutorials (with typical use cases). 
 
 
 
