### 2023-02-20 db09df9
correction in plotting in linux instruction
