### 2023-02-20 #4763 @obertscheiderfhwn
@mrbuche: I was able to reproduce the error in file selection in the GUI on my Ubuntu. Until this will be resolved, I added two different `How to get` sections in the Readme (https://github.com/obertscheiderfhwn/RTMsim#how-to-use---windows and https://github.com/obertscheiderfhwn/RTMsim#how-to-use---linux) where it is also mentioned that the GUI for Linux is still under development.

Running the V&V test cases is possible also from the terminal. I updated the first section on https://obertscheiderfhwn.github.io/RTMsim/build/tutorials/ and also added two sections https://obertscheiderfhwn.github.io/RTMsim/build/tutorials/#Use-cases-Windows and https://obertscheiderfhwn.github.io/RTMsim/build/tutorials/#Use-cases-Linux for the use cases.
