### 2023-02-20 #4763 @obertscheiderfhwn
Addendum: The current version includes a GUI which works also on my Ubuntu. I had to replace the file picking from the package NativeFileDialog by the one from Gtk.
