+++ paper.bib
@@ -46,11 +46,13 @@ volume = {37},
 journal = {Polymer Composites},
 doi = {10.1002/pc.23474}
 }
-@mastersthesis{BSc_ISSE,
-  author  = "Leo Heber",
-  title   = "Reinforcement learning for Resin Transfer Moulding",
-  school  = "Institute for Software and Systems Engineering / University of Augsburg",
-  year    = "in preparation for 2022"
+
+@misc{stieber,
+  title = {Control of {RTM} porcesses via Deep Reinforcement Learning},
+  author = {Stieber, Simon and Heber, Leonard and Obertscheider, Christof and Reif, Wolfgang},
+  howpublished = "www.researchgate.net/publication/367191417\_Control\_of\_RTM\_processes\_through\_Deep\_Reinforcement\_Learning",
+  month="February",
+  year = "2023",
 }
 
 @online{lims,
@@ -79,9 +81,48 @@ author = {Z. Tukovic and H. Jasak},
 keywords = {Collocated finite volume method, Moving polyhedral mesh, Interface tracking, Surface tension, Multiphase fluid flow, PISO algorithm, Rhie–Chow interpolation, OpenFOAM}
 }
 
+@article{doi:10.1137/141000671,
+author = {Bezanson, Jeff and Edelman, Alan and Karpinski, Stefan and Shah, Viral B.},
+title = {Julia: A Fresh Approach to Numerical Computing},
+journal = {SIAM Review},
+volume = {59},
+number = {1},
+pages = {65-98},
+year = {2017},
+doi = {10.1137/141000671}
+}
 
+@book{rudd1997liquid,
+  title={Liquid Moulding Technologies: Resin Transfer Moulding, Structural Reaction Injection Moulding, and Related Processing Techniques},
+  author={Rudd, C.D.},
+  isbn={9780768000160},
+  lccn={97019335},
+  url={https://books.google.at/books?id=yVwE6-3YohcC},
+  year={1997},
+  publisher={Society of Automotive Engineers}
+}
 
+@manual{fluent,
+author = {ANSYS, Inc.},
+title = {Ansys Fluent User's Guide},
+version = {Version 2021 R2}
+}
 
+@book{strong2008fundamentals,
+  title={Fundamentals of Composites Manufacturing, Second Edition: Materials, Methods and Applications},
+  author={Strong, A.B.},
+  isbn={9780872638549},
+  lccn={2007935302},
+  url={https://books.google.at/books?id=aCm9yvodiJcC},
+  year={2008},
+  publisher={Society of Manufacturing Engineers}
+}
 
-
-
+@book{Versteeg,
+	author = {Versteeg, H. K. and Malalasekera, W.},
+	title = {An introduction to computational fluid dynamics :},
+	publisher = {Pearson Education Ltd.,},
+	year = {2007.},
+	address = {Harlow, England ;},
+	edition = {2nd ed.}
+}
