+++ paper.bib
@@ -105,7 +105,8 @@ doi = {10.1137/141000671}
 @manual{fluent,
 author = {ANSYS, Inc.},
 title = {Ansys Fluent User's Guide},
-version = {Version 2021 R2}
+version = {Version 2021 R2},
+year={2021},
 }
 
 @book{strong2008fundamentals,
