+++ paper.md
@@ -25,7 +25,7 @@ bibliography: paper.bib
 RTMsim is a robust, easy-to-use and simple-to-extend software tool, implemented in Julia [@doi:10.1137/141000671] for resin transfer moulding (RTM) filling simulations. A shell mesh, injection pressure, resin viscosity and the parameters describing the preform are required input. The software was validated with results from literature [@rudd1997liquid] and compared with results from well-established RTM filling simulation tools [@myrtm],[@fluent].
 
 # Statement of need
-Resin Transfer Moulding (RTM) is a manufacturing process for producing thin-walled fiber reinforced polymer composites where dry fibers are placed inside a mould and resin is injected under pressure into the fibrous preform. The thickness of the part is much smaller than the overall dimensions $$L \gg H$$ as depicted in \autoref{fig:rtm_pic1}. During mould design, filling simulations can study different manufacturing concepts (i.e. placement of inlet ports and vents) to guarantee complete filling of the part and avoid air entrapment where flow fronts converge. 
+Resin Transfer Moulding (RTM) is a manufacturing process for producing thin-walled fiber reinforced polymer composites where dry fibers are placed inside a mould and resin is injected under pressure into the fibrous preform. The thickness of the part is much smaller than the overall dimensions $$H \ll L$$ as depicted in \autoref{fig:rtm_pic1}. During mould design, filling simulations can study different manufacturing concepts (i.e. placement of inlet ports and vents) to guarantee complete filling of the part and avoid air entrapment where flow fronts converge. 
 
 ![Schematic of a thin-shell RTM geometry for manufacturing a flat plate with linear (a) and radial (b) flow.\label{fig:rtm_pic1}](figures/rtm_pic1.png)
 
@@ -61,30 +61,42 @@ $$ \frac{\partial \rho \mathbf{u}}{\partial t} + \nabla \cdot \left( \rho \mathb
 $$ \frac{\partial \varepsilon c}{\partial t} + \mathbf{u} \cdot \left( \nabla c \right) = 0 $$
 $$ p = \kappa \rho^\gamma $$
 where
+
 - $\rho$ is the mass density of the fluid, 
+
 - $\mathbf{u}$ is the superficial velocity which is related to the physical velocity via the porosity $\varepsilon$ given by the stationary fibrous media, 
+
 - $\mathbf{u}\mathbf{u}$ is the dyadic product, 
+
 - $p$ is the pressure, 
+
 - $-\mu \mathsf{K}^{-1} \mathbf{u}$ is the pressure loss from flow through a porous medium according to Darcy's law with dynamic viscosity $\mu$ and permeability tensor $\mathsf{K}$, 
+
 - $c$ is the continuous fraction function with $0 \leq c \leq 1$, $c=0$ if empty and $c=1$ if completely filled, 
+
 - $\gamma$ is the adiabatic index, and 
+
 - $\kappa=p_{\rm ref}/\rho_{\rm ref}^\gamma$ is a constant determined from pressure and mass density reference values.
 
+Since the filling is a result of the pressure difference between injection pressure and initial cavity pressure, gauge pressures are used for the simulation. The initial pressure in the cavity is set equal to zero and the injection pressure is set equal to the pressure difference. This normalization results in a filling which is independent of the pressure level and only depends on the pressure difference. 
+
 It is assumed that no cross-thickness flow takes place. Consequently, the fluid domain can be meshed with only one cell through the thickness or with a shell mesh on the mid-surface, see \autoref{fig:theory_001}(a) and (b). The red inlet and optional outlet ports are also part of the mesh, see \autoref{fig:theory_001}&#40;c). 
+
[…]
