+++ paper.md
@@ -80,27 +80,15 @@ where
 
 Since the filling is a result of the pressure difference between injection pressure and initial cavity pressure, gauge pressures are used for the simulation. The initial pressure in the cavity is set equal to zero and the injection pressure is set equal to the pressure difference. This normalization results in a filling which is independent of the pressure level and only depends on the pressure difference. 
 
-It is assumed that no cross-thickness flow takes place. Consequently, the fluid domain can be meshed with only one cell through the thickness or with a shell mesh on the mid-surface, see \autoref{fig:theory_001}(a) and (b). The red inlet and optional outlet ports are also part of the mesh, see \autoref{fig:theory_001}&#40;c). 
+In order to numerically solve the flow model the computational domain (time and space) is discretized. The temporal domain (i.e. the simulation time) is split into a finite number of time steps where values for the physical quantities on the spatial domain are calculated in a time-marching manner. The spatial domain (i.e. the flow volume) is split into wedge-type cells. Since the walls are assumed to be slip boundaries only one cell is used through the thickness and the cells are bounded by three control surfaces only. Therefore, the spatial domain is defined on the part’s mid-surface and the local thickness of the flow volume is a property of the cell. The mid-surface model can be curved and cells can have edges where more than two cells are connected to each other such as for handling T-junctions. The mid-surface model is divided into a finite number of triangular control areas. The control areas cover the spatial domain completely without overlapping. The cell-centered method where the nodes are placed at the centroid of the control volume is used. 
 
-![Volume mesh (a) and shell mesh (b) and domain with highlighted inlet region &#40;c) for the simulation of a radial flow experiment. \label{fig:theory_001}](figures/theory_001.png)
+Since a shell mesh is used in the simulation tool, the conservation laws must be solved on a shell mesh using a generalization of the so-called finite area method [@tukovic]. The governing equations are discretized for every cell of the shell mesh.
 
-In general, flow through curved cavity with junctions must be solved, see \autoref{fig:theory_002}(a). The discretized versions of the continuum, momentum and volume-of-fluid equations are solved on the flattened geometry. The neighbouring cells are rotated about the common edge to lie in a plane, see \autoref{fig:theory_002}(b) and &#40;c).
-
-![Annulus filler mesh with T-sections and curved regions. \label{fig:theory_002}](figures/theory_002.png)
-
-Since the flow is described on the mid-surface, the velocity vector has only two non-zero components u and v described in a local coordinate system. The local coordinate system is the result of a projection of a reference vector. The reference vector is chosen such that the local x- and y-directions correspond with the first and second principal permeability directions. 
-
-The four conservation laws for mass density, velocity components u and v in the local cell coordinate system and the fluid fraction describe the flow in the locally flattened geometry, see \autoref{fig:theory_003}(a). A coordinate transformation for the velocity components must be performed for every neighbouring cell, see \autoref{fig:theory_003}(b).
+the velocity vector has only two non-zero components u and v described in a local coordinate system. The local coordinate system is the result of a projection of a reference vector. The reference vector is chosen such that the local x- and y-directions correspond with the first and second principal permeability directions. For the evaluation of the numerical flux functions and the pressure gradient, the neighbouring triangular cells are rotated about the common edge to lie in the plane of the considered cell and a coordinate transformation for the velocity components must be performed for every neighbouring cell. This is illustrated in \autoref{fig:theory_003}(a) and (b). 
 
 ![Sketches illustrating the flattened mesh. \label{fig:theory_005}](figures/theory_005.png)
 
[…]
