+++ paper.bib
@@ -121,9 +121,9 @@ year={2021},
 
 @book{Versteeg,
 	author = {Versteeg, H. K. and Malalasekera, W.},
-	title = {An introduction to computational fluid dynamics :},
-	publisher = {Pearson Education Ltd.,},
-	year = {2007.},
-	address = {Harlow, England ;},
+	title = {An introduction to computational fluid dynamics},
+	publisher = {Pearson Education Ltd.},
+	year = {2007},
+	address = {Harlow, England},
 	edition = {2nd ed.}
 }
