+++ paper.md
@@ -82,11 +82,9 @@ Since the filling is a result of the pressure difference between injection press
 
 In order to numerically solve the flow model the computational domain (time and space) is discretized. The temporal domain (i.e. the simulation time) is split into a finite number of time steps where values for the physical quantities on the spatial domain are calculated in a time-marching manner. The spatial domain (i.e. the flow volume) is split into wedge-type cells. Since the walls are assumed to be slip boundaries only one cell is used through the thickness and the cells are bounded by three control surfaces only. Therefore, the spatial domain is defined on the part’s mid-surface and the local thickness of the flow volume is a property of the cell. The mid-surface model can be curved and cells can have edges where more than two cells are connected to each other such as for handling T-junctions. The mid-surface model is divided into a finite number of triangular control areas. The control areas cover the spatial domain completely without overlapping. The cell-centered method where the nodes are placed at the centroid of the control volume is used. 
 
-Since a shell mesh is used in the simulation tool, the conservation laws must be solved on a shell mesh using a generalization of the so-called finite area method [@tukovic]. The governing equations are discretized for every cell of the shell mesh.
+Since a shell mesh is used in the simulation tool, the conservation laws must be solved on a shell mesh using a generalization of the so-called finite area method [@tukovic]. The governing equations are discretized for every cell of the shell mesh. Consequently, the velocity vector has only two non-zero components $u$ and $v$ described in a local coordinate system. The local coordinate system is the result of a projection of a reference vector. The reference vector is chosen such that the local x- and y-directions correspond with the first and second principal permeability directions. For the evaluation of the numerical flux functions and the pressure gradient, the neighbouring triangular cells are rotated about the common edge to lie in the plane of the considered cell and a coordinate transformation for the velocity components must be performed for every neighbouring cell. This is illustrated in \autoref{fig:theory_005}. 
 
-the velocity vector has only two non-zero components u and v described in a local coordinate system. The local coordinate system is the result of a projection of a reference vector. The reference vector is chosen such that the local x- and y-directions correspond with the first and second principal permeability directions. For the evaluation of the numerical flux functions and the pressure gradient, the neighbouring triangular cells are rotated about the common edge to lie in the plane of the considered cell and a coordinate transformation for the velocity components must be performed for every neighbouring cell. This is illustrated in \autoref{fig:theory_003}(a) and (b). 
-
-![Sketches illustrating the flattened mesh. \label{fig:theory_005}](figures/theory_005.png)
+![Flattended geometry (a) with local coordinate systems of the neighbouring cells (b). \label{fig:theory_005}](figures/theory_005.png)
 
 \autoref{fig:mesh_pic1} shows the shell mesh which is used for two out of three test cases, which are reported in the subsequent section.
 
@@ -117,7 +115,7 @@ A follow-up paper will discuss the physical model in detail and show the results
 
 # Outlook
 The existing software packages for RTM mould filling simulations perform well for most scientific and engineering applications, but
-there are applications where a new adaptable software package are required. For example, a transient pressure drop in the cavity after closing a valve in the hose between pressure tank and test rig can be studied with RTMsim after some extensions. Furthermore, the parts of RTMsim package can be integrated in separate codes. For example, a customized lightweight version of the solver was already used as training environment for Reinforcement learning (RL) algorithms to keep the flow front straight in a rectangular flow with reinforcement patches and three inlet ports [@stieber].
[…]
