+++ paper.md
@@ -82,7 +82,7 @@ Since the filling is a result of the pressure difference between injection press
 
 In order to numerically solve the flow model the computational domain (time and space) is discretized. The temporal domain (i.e. the simulation time) is split into a finite number of time steps where values for the physical quantities on the spatial domain are calculated in a time-marching manner. The spatial domain (i.e. the flow volume) is split into wedge-type cells. Since the walls are assumed to be slip boundaries only one cell is used through the thickness and the cells are bounded by three control surfaces only. Therefore, the spatial domain is defined on the part’s mid-surface and the local thickness of the flow volume is a property of the cell. The mid-surface model can be curved and cells can have edges where more than two cells are connected to each other such as for handling T-junctions. The mid-surface model is divided into a finite number of triangular control areas. The control areas cover the spatial domain completely without overlapping. The cell-centered method where the nodes are placed at the centroid of the control volume is used. 
 
-Since a shell mesh is used in the simulation tool, the conservation laws must be solved on a shell mesh using a generalization of the so-called finite area method [@tukovic]. The governing equations are discretized for every cell of the shell mesh. Consequently, the velocity vector has only two non-zero components $u$ and $v$ described in a local coordinate system. The local coordinate system is the result of a projection of a reference vector. The reference vector is chosen such that the local x- and y-directions correspond with the first and second principal permeability directions. For the evaluation of the numerical flux functions and the pressure gradient, the neighbouring triangular cells are rotated about the common edge to lie in the plane of the considered cell and a coordinate transformation for the velocity components must be performed for every neighbouring cell. This is illustrated in \autoref{fig:theory_005}. 
+Since a shell mesh is used in the simulation tool, the conservation laws must be solved on a shell mesh using a generalization of the so-called finite area method [@tukovic]. The governing equations are discretized for every cell of the shell mesh. Consequently, the velocity vector has only two non-zero components $u$ and $v$ described in a local coordinate system. The local coordinate system is the result of a projection of a reference vector. The reference vector is chosen such that the local $x$- and $y$-directions correspond with the first and second principal permeability directions. For the evaluation of the numerical flux functions and the pressure gradient, the neighbouring triangular cells are rotated about the common edge to lie in the plane of the considered cell and a coordinate transformation for the velocity components must be performed for every neighbouring cell. This is illustrated in \autoref{fig:theory_005}. 
 
 ![Flattended geometry (a) with local coordinate systems of the neighbouring cells (b). \label{fig:theory_005}](figures/theory_005.png)
 
