+++ paper.md
@@ -22,7 +22,7 @@ bibliography: paper.bib
 ---
 
 # Summary
-RTMsim is a robust, easy-to-use and simple-to-extend software tool, implemented in Julia [@doi:10.1137/141000671] for resin transfer moulding (RTM) filling simulations. A shell mesh, injection pressure, resin viscosity and the parameters describing the preform are required input. The software was validated with results from literature [@rudd1997liquid] and compared with results from well-established RTM filling simulation tools [@myrtm],[@fluent].
+RTMsim is a robust, easy-to-use and simple-to-extend software tool, implemented in Julia [@doi:10.1137/141000671] for resin transfer moulding (RTM) filling simulations. A shell mesh, injection pressure, resin viscosity and the parameters describing the preform are required input. The software was validated with results from literature [@advani] and compared with results from well-established RTM filling simulation tools [@myrtm],[@fluent].
 
 # Statement of need
 Resin Transfer Moulding (RTM) is a manufacturing process for producing thin-walled fiber reinforced polymer composites where dry fibers are placed inside a mould and resin is injected under pressure into the fibrous preform. The thickness of the part is much smaller than the overall dimensions $$H \ll L$$ as depicted in \autoref{fig:rtm_pic1}. During mould design, filling simulations can study different manufacturing concepts (i.e. placement of inlet ports and vents) to guarantee complete filling of the part and avoid air entrapment where flow fronts converge. 
@@ -96,7 +96,7 @@ Since a shell mesh is used in the simulation tool, the conservation laws must be
 
 Different test cases are available, successfully validating the Julia implementation of the RTM filling model. Test cases 1 and 2 are so-called numerical permeameter experiments, see e.g. [@Fauster.2019]. The third test case is a filling study of a typical composite part which is manufactured using RTM, see [@barandun]. The testcases are:
 
-1. Validation of the software for radial flow with isotropic in-plane permeablity: The simulated flow front position after 200 s is compared with the calculated flow front position from literature [@rudd1997liquid]. The preform size is $0.6 \times 0.6 \times 0.003$ m$^3$ with a central injection gate with 0.02 m diameter. The preform has porosity $0.70$ and isotropic permeability $3.0 \cdot 10^{−10}$ m$^2$. Resin with dynamic viscosity 0.06 Pas is injected with 35000 Pa pressure.
+1. Validation of the software for radial flow with isotropic in-plane permeablity: The simulated flow front position after 200 s is compared with the calculated flow front position from literature [@advani]. The preform size is $0.6 \times 0.6 \times 0.003$ m$^3$ with a central injection gate with 0.02 m diameter. The preform has porosity $0.70$ and isotropic permeability $3.0 \cdot 10^{−10}$ m$^2$. Resin with dynamic viscosity 0.06 Pas is injected with 35000 Pa pressure.
 
 2. Verification of the software for radial flow with tilted orthotropic in-plane permeablity: The simulated tilted elliptical flow front is analysed and the calculated orthotropic permeablity is compared with the input (permeability $k_1$ and $k_2$, angle $\theta$ of the first principal direction to the horizontal). The preform size is $0.6 \times 0.6 \times 0.003$ m$^3$ with a central injection gate with 0.02 m diameter. The preform has porosity $0.70$ and permeability $3.0 \cdot 10^{−10}$ m$^2$ in first principal direction $30^\circ$ to the horizontal and permeability $1.5 \cdot 10^{−10}$ m$^2$ in second principal direction. Resin with dynamic viscosity 0.06 Pas is injected with 35000 Pa pressure. 
 
