### 2023-02-21 obertscheiderfhwn/RTMsim#2 @obertscheiderfhwn
Dear @Zitzeronion, I updated the the paper according to (hopefully all) your suggestions. Please have a look at the new version:
[paper_20230221.pdf](https://github.com/obertscheiderfhwn/RTMsim/files/10797759/paper_20230221.pdf)
