### 2023-02-21 #4763 @obertscheiderfhwn
@Zitzeronion and @mrbuche: Paper updated, see https://github.com/obertscheiderfhwn/RTMsim/issues/2
