### 2023-02-22 obertscheiderfhwn/RTMsim#2 @Zitzeronion
Thanks for the additions to the paper, especially the modeling equations.
There is one more improve for the overall readability of the paper, namely the reduction of size of Fig. 4-5.
Apart from that green light for the paper from my side.
