### 2023-02-23 obertscheiderfhwn/RTMsim#2 @obertscheiderfhwn
Green light refers to the small green circles in Fig. 4 and 5?
