### 2023-02-23 obertscheiderfhwn/RTMsim#2 @Zitzeronion
Figure 4-5 fill a whole page each, I think the paper would not suffer if you downscale these plots a little and put the table (which is now part of Fig 4 -> Fig 4b) in the text as a markdown table.

No, the green light refers to the state of the paper and that I am overall satisfied (given this minor stylistic thing are changed) 👍
