+++ paper.md
@@ -102,16 +102,26 @@ Different test cases are available, successfully validating the Julia implementa
 
 3. The filling of a complex annulus filler-like part is simulated with different software tools. The bounding box of the part is approximately $0.4 \times 0.15 \times 0.1$ m$^3$. The flow front positions at different time instances simulated with Ansys Fluent and RTMsim are compared in case 3.1. The filling patterns for same filling degrees simulated with myRTM and RTMsim are compared in case 3.2.
 
-\autoref{fig:pic1}, \autoref{fig:pic2} and \autoref{fig:pic3} show the binary filling fraction for cases 1, 2 and 3. For every cell $i$ the fraction function is either $0$ or $1$, depending on the filling fraction being $\leq$ or $>$ than a threshold value. The threshold for the volume-of-fluid method is determined from the permeameter simulation with isotropic permeabilities (case 1). The circular and elliptical flow fronts for cases 1 and 2 are calculated from the simulated filling fraction contour plots with an image processing algorithm [@Fauster.2019]. For case 1 also a comparison with results from an analytical formula [@advani] is shown. 
+\autoref{fig:pic1}, \autoref{fig:pic2} and \autoref{fig:pic3} show the binary filling fraction for cases 1, 2 and 3. For every cell $i$ the fraction function is either $0$ or $1$, depending on the filling fraction being $\leq$ or $>$ than a threshold value. The threshold for the volume-of-fluid method is determined from the permeameter simulation with isotropic permeabilities (case 1). The circular and elliptical flow fronts for cases 1 and 2 are calculated from the simulated filling fraction contour plots with an image processing algorithm [@Fauster.2019]. 
 
-A follow-up paper will discuss the physical model in detail and show the results of a validation with experimental data from a radial flow experiment with two patches with different in-plane permeability and porosity levels.
+![Simulated flow front for case 1 after 200 s for a coarse mesh (a) and a fine mesh (b). The left pictures show the simulation results and the right pictures show the calculated flow front positions. Simulation input for the porosity was $0.70$ and for the isotropic permeability was $3.0 \cdot 10^{−10}$ m$^2$. The flow front position is similar but the shape is no smooth circle for the coarse mesh.\label{fig:pic1}](figures/validation_pic1_mod.png)
 
-![Simulated flow front for case 1 after 200 s for a coarse and a fine mesh and flow front positions at different time instances calculated with an analytical formula and calculated from the simulated flow front with an image processing algorithm for a coarse and a fine mesh. The values for the fine mesh agree well with the results from the analytical formula. The values for the coarse mesh show an error of $\leq 15$% which decreases significantly in the course of time but the shape is no smooth circle.\label{fig:pic1}](figures/validation_pic1.png)
+For case 1 also a comparison with results from an analytical formula [@advani] is performed. The following table shows the flow front positions as radial distance $r$ from the center at different time instances $t$ calculated with an analytical formula and calculated from the simulated flow front with an image processing algorithm for a coarse and a fine mesh.
 
[…]
