+++ paper.md
@@ -119,7 +119,7 @@ The values for the fine mesh agree well with the results from the analytical for
 
 ![Simulated flow front for case 2 after 200 s for a coarse mesh (a) and a fine mesh (b). The left pictures show the simulation results and the right pictures show the calculated flow front positions. Simulation input for the porosity was $0.70$ and for the permeability was $3.0 \cdot 10^{−10}$ m$^2$ in first principal direction $30^\circ$ to the horizontal and $1.5 \cdot 10^{−10}$ m$^2$ in second principal direction. An algorithm for determining the flow front position from an optical permeameter was adapted to calculate the permeablity. The results are $2.97 \cdot 10^{−10}$ m$^2$, $1.45 \cdot 10^{−10}$ m$^2$ and angle $35^\circ$ for the fine mesh and $2.48 \cdot 10^{−10}$ m$^2$, $1.18 \cdot 10^{−10}$ m$^2$ and $35^\circ$ for the coarse mesh. For the fine mesh (2198 cells for a domain with 600 × 600 mm), this reverse engineering shows very good agreement of the calculated permeability values with the values used as simulation input. For the coarse mesh (588 cells) the agreement is still acceptable.\label{fig:pic2}](figures/validation_pic2_mod.png)
 
-![Results for test case 3. The results for the new simulation tool with the fine mesh are in good agreement with the ANSYS Fluent results for both filling pattern and filling time. myRTM shows similar filling patterns for same filling degrees as the new simulation tool with the coarse mesh. This comparison shows that a coarse mesh can be used for filling simulations where only the flow front propagation is investigated. If the filling time is important a mesh refinement study must be performed. \label{fig:pic3}](figures/validation_pic3.png)
+![Results for test case 3. The results for the new simulation tool with the fine mesh are in good agreement with the ANSYS Fluent results for both filling pattern and filling time. myRTM shows similar filling patterns for same filling degrees as the new simulation tool with the coarse mesh. This comparison shows that a coarse mesh can be used for filling simulations where only the flow front propagation is investigated. If the filling time is important a mesh refinement study must be performed. \label{fig:pic3}](figures/validation_pic3_mod.png)
 
 A follow-up paper will discuss the physical model in detail and show the results of a validation with experimental data from a radial flow experiment with two patches with different in-plane permeability and porosity levels.
 
