### 2023-02-23 obertscheiderfhwn/RTMsim#2 @obertscheiderfhwn
Changes implemented and pdf created on https://github.com/openjournals/joss-reviews/issues/4763
