### 2023-02-26 obertscheiderfhwn/RTMsim#4 @Zitzeronion
Docstrings for functions
Sorry @obertscheiderfhwn for taking my time on the package 😅
There are some things that I would like you implement in order to check of the remain boxes of the [Joss submission](https://github.com/openjournals/joss-reviews/issues/4763)

The functions:
- `create_coordinate_system( )`
- `assign_parameters( )`
- `create_faces( )`
- `gui( )`
don't have proper docstrings, please add at least an argument section to the docstrings.
They don't need to be as precise as outlined in the [official Julia docs](https://docs.julialang.org/en/v1/manual/documentation/), but some more documentation would be helpful.
