### 2023-02-26 obertscheiderfhwn/RTMsim#5 @Zitzeronion
Unit test for `numerical_flux_function`
Similar to [issue](https://github.com/obertscheiderfhwn/RTMsim/issues/4) one the remaining boxes for the [Joss paper](https://github.com/openjournals/joss-reviews/issues/4763) covers testing.

I do like the integration test that comes with the package.
For any further development unit tests could help to reduce the time spend debugging.
There are two functions that would benefit from an unit test:
-`numerical_flux_function( )`
-`numerical_flux_function_boundary( )`

If you find yourself having trouble implementing an unit test let me know and I can write a PR.
