+++ inputfiles/input_case4.txt
@@ -1,5 +1,5 @@
 1    #i_model 
-meshfiles\\mesh_permeameter5.bdf    #meshfilename 
+meshfiles/mesh_permeameter5.bdf    #meshfilename 
 160    #tmax
 1.01325e5 1.225 1.4 0.071    #p_ref rho_ref gamma mu_resin_val 
 1.91e5 1.0e5    #p_a_val p_init_val 
