### 2023-02-28 48cbf2c
`p_init=p_init-p_init+p_eps` -> `p_init = p_eps`
Less complicated :)
