### 2023-03-01 #4763 @Zitzeronion
Hi @Fei-Tao, I'm almost done with my review.
@obertscheiderfhwn did a fantastic job improving the software and the paper so far.

There are two things that I want @obertscheiderfhwn to add before I can check off the last boxes:
1. Tests, see [issue 4](https://github.com/obertscheiderfhwn/RTMsim/issues/5)
2. Homogenization of docstrings, see [issue](https://github.com/obertscheiderfhwn/RTMsim/issues/4)

Minor things that I think would improve the package even more, but this is personal preference and leave the decision to you @Fei-Tao:
- Trailing semicolons are not needed in Julia
- Breaking code up into logical bits, CFD one file, plotting one file, gui one file.
- Typed numbers, AFAIK the default in Julia is Float64, there is no need to have `Float64(0.0)`
- I really wish there were structs in rtmsim, I think they would help massively with future proofing the package.

The scientific goal of the package is met and my criticism is mainly due to the possible improvements to the codebase.
