+++ src/core.jl
@@ -1,0 +1,742 @@
+"""
+    rtmsim_rev1(param)
+
+RTMsim solver with the following main steps:
+- Simulation initialization
+- Read mesh file and prepare patches  
+- Find neighbouring cells
+- Assign parameters to cells
+- Create local cell coordinate systems
+- Calculate initial time step
+- Array initialization
+- Define simulation time and intermediate output times
+- Boundary conditions
+- (Optional initialization if `i_model=2,3,..`)
+- Time evolution (for loops over all indices inside a while loop for time evolution)
+    - Calculation of correction factors for cell thickness, porosity, permeability, viscosity
+    - Pressure gradient calculation
+    - Numerical flux function calculation
+    - Update of rho, u, v, gamma and p according to conservation laws and equation of state
+    - Boundary conditions
+    - Prepare arrays for next time step
+    - Saving of intermediate data
+    - (Opional time marching etc. for i_model=2,3,...)
+    - Calculation of adaptive time step 
+
+Data structure `param` with arguments:
+- i_model :: Int64 =1
+- meshfilename :: String ="input.txt"
+- tmax :: Float64 =200
+- p_ref :: Float64 =1.01325e5
+- rho_ref :: Float64 =1.225
+- gamma :: Float64 =1.4
+- mu_resin_val :: Float64 =0.06
+- p_a_val :: Float64 =1.35e5
+- p_init_val :: Float64 =1.0e5      
+- t_val :: Float64 =3e-3
+- porosity_val :: Float64 =0.7
+- K_val :: Float64 =3e-10
+- alpha_val :: Float64 =1.0
+- refdir1_val :: Float64 =1.0
+- refdir2_val :: Float64 =0.0
+- refdir3_val :: Float64 =0.0
+- t1_val :: Float64 =3e-3
+- porosity1_val :: Float64 =0.7
+- K1_val :: Float64 =3e-10
+- alpha1_val :: Float64 =1.0
+- refdir11_val :: Float64 =1.0
+- refdir21_val :: Float64 =0.0
+- refdir31_val :: Float64 =0.0
+- t2_val :: Float64 =3e-3
+- porosity2_val :: Float64 =0.7
+- K2_val :: Float64 =3e-10
+- alpha2_val :: Float64 =1.0
+- refdir12_val :: Float64 =1.0
+- refdir22_val :: Float64 =0.0
+- refdir32_val :: Float64 =0.0
+- t3_val :: Float64 =3e-3
+- porosity3_val :: Float64 =0.7
+- K3_val :: Float64 =3e-10
+- alpha3_val :: Float64 =1.0
+- refdir13_val :: Float64 =0.0
+- refdir23_val :: Float64 =0.0
+- refdir33_val :: Float64 =0.0
+- t4_val :: Float64 =3e-3
+- porosity4_val :: Float64 =0.7
+- K4_val :: Float64 =3e-10
+- alpha4_val :: Float64 =1.0
+- refdir14_val :: Float64 =1.0
+- refdir24_val :: Float64 =0.0
+- refdir34_val :: Float64 =0.0
+- patchtype1val :: Int64 =1
+- patchtype2val :: Int64 =0
+- patchtype3val :: Int64 =0
+- patchtype4val :: Int64 =0
+- i_restart :: Int64 =0
+- restartfilename :: String ="results.jld2"
+- i_interactive :: Int64 =0
+- r_p :: Float64 =0.01
+- n_pics :: Int64 =16
+
+Meaning of the variables:
+- `i_model`: Identifier for physical model (Default value is 1)
+- `meshfilename`: Mesh filename.
+- `tmax`: Maximum simulation time.
+- `p_ref rho_ref gamma mu_resin_val`: Parameters for the adiabatic equation of state and dynamic viscosity of resin used in the Darcy term.
[…]
