### 2023-03-03 76f8e77
Suggestions by @Zitzeronion
