+++ src/tools.jl
@@ -1,19 +1,27 @@
 """
-    function create_faces(cellgridid, N, maxnumberofneighbours)
+    function create_faces(input_struct)
 
 Find the set with the IDs of the neighbouring cells and identify wall cells.
 
-Arguments:
-- cellgridid :: Array{Int,2}
-- N :: Int
-- maxnumberofneighbours :: Int
+Arguments: Data structure with
+- `cellgridid :: Array{Int,2}`
+- `N :: Int`
+- `maxnumberofneighbours :: Int`
 
 Meaning of the arguments:
 - `cellgridid`: The i-th line contains the three IDs of the nodes which form the cell
 - `N`: Number of cells.
 - `maxnumberofneighbours`: Number of columns of array `cellneighboursarray`. Default value is `10`. If more cell neighbours in the mesh, an error occurs and this value must be increased. 
+
+Return: Data structure with
+- `faces :: Array{Int,2}`
+- `cellneighboursarray :: Array{Int,2}`
+- `celltype :: Vector{Int}`
 """
-function create_faces(cellgridid, N, maxnumberofneighbours)
+function create_faces(input_struct)
+    cellgridid=input_struct.cellgridid
+    N=input_struct.N
+    maxnumberofneighbours=input_struct.maxnumberofneighbours
     celltype=Vector{Int64}(undef, N)
     for i in 1:N
         celltype[i]=1
@@ -83,7 +91,8 @@ function create_faces(cellgridid, N, maxnumberofneighbours)
         end
     end
 
-    return faces, cellneighboursarray, celltype
+    return_struct=rtmsim.return_args_create_faces(faces, cellneighboursarray, celltype)
+    return return_struct
 end
 
 
@@ -99,19 +108,29 @@ end
 
 
 """
-    function assign_parameters(i_interactive,celltype,patchparameters0,patchparameters1,patchparameters2,patchparameters3,patchparameters4,patchtype1val,patchtype2val,patchtype3val,patchtype4val,patchids1,patchids2,patchids3,patchids4,inletpatchids,mu_resin_val,N)
+    function assign_parameters(input_struct)
 
 Assign properties to cells.
 
-Arguments:
-- i_interactive :: Int
-- celltype :: Vector{Int}
-- patchparameters0,patchparameters1,patchparameters2,patchparameters3,patchparameters4 :: Vector{Float}
-- patchtype1val,patchtype2val,patchtype3val,patchtype4val :: 
-- patchids1,patchids2,patchids3,patchids4 :: Int
-- inletpatchids :: Vector{Int}
-- mu_resin_val :: Float
-- N :: Int
+Arguments: Data structure with
+- `i_interactive :: Int`
+- `celltype :: Vector{Int}`
+- `patchparameters0 :: Vector{Float}`
+- `patchparameters1 :: Vector{Float}`
+- `patchparameters2 :: Vector{Float}`
+- `patchparameters3 :: Vector{Float}`
+- `patchparameters4 :: Vector{Float}`
+- `patchtype1val :: Int`
+- `patchtype2val :: Int`
+- `patchtype3val :: Int`
+- `patchtype4val :: Int`
+- `patchids1 :: Vector{Int}`
+- `patchids2 :: Vector{Int}`
+- `patchids3 :: Vector{Int}`
+- `patchids4 :: Vector{Int}`
+- `inletpatchids :: Vector{Int}`
+- `mu_resin_val :: Float`
+- `N :: Int`
 
 Meaning of the arguments:
 - `i_interactive`: 1..if pressure inlet cells are selected manually, 0..else
@@ -122,8 +141,36 @@ Meaning of the arguments:
 - `inletpatchids` :: Vector with the IDs of the cells which are inlet cells.
 - `mu_resin_val` :: Kinematic viscosity value.
[…]
