+++ src/num.jl
@@ -133,8 +133,12 @@ Return: Data structure with
 - `F_gamma_num1_add :: Float`
 
 Unit tests:
-- Comparison with flux function for one-dimensional case without upwinding for the continuity equation and with upwinding in the momentum equation, i.e. F_rho=(rho*u)_withoutupwinding=0.5*(1.0+1.0)*0.5*(1.2+0.4)=0.8 and F_u=(rho*u)_withoutupwinding*u_withupwinding=0.8*1.2=0.96 in density equation without upwinding and rho*u without upwinding. rho=1.0 in considered and neighbouring cell, u=1.2 in considered cell and u=0.4 in neighbouring cell, v=0 and gamma=0 in considered and neighbouring cell. `input_struct=rtmsim.input_args_flux(1,[1.0; 1.2; 0.0; 0.0],[1.0; 0.4; 0.0; 0.0],[1.0;0.0;1.0]);return_struct=rtmsim.numerical_flux_function(input_struct);println("F_rho="*string(return_struct.F_rho_num_add)*", F_u="*string(return_struct.F_u_num_add))` with result `F_rho=0.8, F_u=0.96`
-- General case with [rho,u,v,gamma]=[1.225,1.2,0.4,0.9] in the considered cell and [rho,u,v,gamma]=[1.225,1.2,0.4,0.9] in the neighbouring cell and outwards pointing cell normal vector [1/sqrt(2),1/sqrt(2)] and boundary face area 1.0. `input_struct=rtmsim.input_args_flux(1,[1.225; 1.2; 0.4; 0.9],[1.0; 0.4; 1.2; 0.1],[1/sqrt(2);1/sqrt(2);1.0]);return_struct=rtmsim.numerical_flux_function(input_struct);println("F_rho="*string(return_struct.F_rho_num_add)*", F_u="*string(return_struct.F_u_num_add)*", F_v="*string(return_struct.F_v_num_add)*", F_gamma="*string(return_struct.F_gamma_num_add)*", F_gamma_num1="*string(return_struct.F_gamma_num1_add))` with result `F_rho=1.2586500705120547, F_u=1.5103800846144655, F_v=0.5034600282048219, F_gamma=1.0182337649086284, F_gamma_num1=1.131370849898476` 
+
+```input_struct=rtmsim.input_args_flux(1,[1.0; 1.2; 0.0; 0.0],[1.0; 0.4; 0.0; 0.0],[1.0;0.0;1.0]);return_struct=rtmsim.numerical_flux_function(input_struct);println("F_rho="*string(return_struct.F_rho_num_add)*", F_u="*string(return_struct.F_u_num_add))``` 
+- Comparison with flux function for one-dimensional case without upwinding for the continuity equation `F_rho=rho*u=0.5*(1.0+1.0)*0.5*(1.2+0.4)=0.8` without upwinding and and `F_u=(rho*u)*(u)=0.8*1.2=0.96` with upwinding in the first factor and without upwinding in the second. Mass density `rho=1.0` in the considered and neighbouring cells, `x` velocity `u=1.2` in considered cell and `u=0.4` in neighbouring cell, `y` velocity `v=0` and fluid fraction `gamma=0` in considered and neighbouring cells. Result is `F_rho=0.8, F_u=0.96`. 
[…]
