### 2023-03-05 #4763 @obertscheiderfhwn
@Zitzeronion and @Fei-Tao, I implemented the minor things and worked on the issues mentioned on the repository. Please have a look and let me know if additional work is required. Thanks in advance!
