### 2023-03-05 #4763 @Zitzeronion
@obertscheiderfhwn the latest changes did cross of the minor things I suggested. One point that you may want to look into in the future is the mutability of the structs. As far as I can tell you have all of them mutable, which is fine, but it could be that mutability comes at a performance cost (if performance is ever an issue).

I'm looking forward to your changes concerning the issues 👍
