### 2023-03-06 bc55066
Docstrings and unit tests
