### 2023-03-06 obertscheiderfhwn/RTMsim#4 @obertscheiderfhwn
Docstrings added for the four functions:
https://obertscheiderfhwn.github.io/RTMsim/build/functions/#rtmsim.create_coordinate_systems-Tuple{Any}
https://obertscheiderfhwn.github.io/RTMsim/build/functions/#rtmsim.assign_parameters-Tuple{Any}
https://obertscheiderfhwn.github.io/RTMsim/build/functions/#rtmsim.create_faces-Tuple{Any}
https://obertscheiderfhwn.github.io/RTMsim/build/functions/#rtmsim.gui-Tuple{}
