### 2023-03-06 obertscheiderfhwn/RTMsim#5 @obertscheiderfhwn
Unit tests added for the two functions:
https://obertscheiderfhwn.github.io/RTMsim/build/functions/#rtmsim.numerical_flux_function-Tuple{Any}
https://obertscheiderfhwn.github.io/RTMsim/build/functions/#rtmsim.numerical_flux_function_boundary-Tuple{Any}

Hope this is what you meant.
