### 2023-03-06 #4763 @obertscheiderfhwn
@Zitzeronion, I tried to solve the open issues https://github.com/obertscheiderfhwn/RTMsim/issues/5 and https://github.com/obertscheiderfhwn/RTMsim/issues/4
