### 2023-03-07 obertscheiderfhwn/RTMsim#4 @Zitzeronion
Thank you for the additional work on the documentation of the functions. This closes the docstring issue I raised for [Joss](https://github.com/openjournals/joss-reviews/issues/4763)
