### 2023-03-07 obertscheiderfhwn/RTMsim#5 @Zitzeronion
That is a good starting point.
Having tests in the docstrings is of course helpful, but I was looking for tests in the runtest file.
Imagine I want to work on `numerical_flux_function` and have some ideas.
I implement changes and call `julia> ] test` to see if I have broken something.
Right now I could break something and the test would work, because in your runtest file contains no `@test` statement.

Could you take the docstring tests and put them into the runtest file?
This is generally a good idea, because within your CI (continuous integration) file (the `.yaml` file in the github folder) you can automate testing and be aware about breaking changes.
And as a user and contributor I can simply call `julia> ] test` to check if my code works or breaks something.
