+++ .github/workflows/runtests.yml
@@ -7,9 +7,11 @@ jobs:
       matrix:
         julia-version: ['1.8']
         julia-arch: [x64]
-        os: [windows-latest]
+        os: [ubuntu-latest]
     steps:
       - uses: actions/checkout@v2
+      - uses: openrndr/setup-opengl@v1.1
+      - run: xvfb-run glxinfo
       - uses: julia-actions/setup-julia@latest
         with:
           version: ${{ matrix.julia-version }}
