### 2023-03-07 obertscheiderfhwn/RTMsim#5 @obertscheiderfhwn
The runtests.jl now includes unit tests for the numerical flux functions, running a typical simulation and creating a contour plot from the simulation results.

Automated testing with GitHub actions was not so easy to implement since rtmsim is using the package GLMakie with windows on a headless environment. But I copied relevant section from GLMakie's ci.yml file to run the tests.

Successfully passed a test:
https://github.com/obertscheiderfhwn/RTMsim/actions/runs/4355606962/jobs/7612454744
