### 2023-03-07 obertscheiderfhwn/RTMsim#5 @Zitzeronion
Well done, but that was not my intention and there is still no test.
Sorry for being not clear in my description, but try to have a look at [DataFrames.jl](https://github.com/JuliaData/DataFrames.jl) [test suit](https://github.com/JuliaData/DataFrames.jl/blob/main/test/index.jl) for a clearer understanding.

Basically the idea is that you execute a mathematical function, say y(x) = x^2, so you know that for x = 3 the function should return 9. In `runtests.jl` you can do the same, by knowing the input you know the output.
In file square.jl
```julia
function y(x)
  return x^2
end
```
and in `runtest.jl` of this package you can do
```julia
@test y(3) == 9
```
Now, if `y(x)` gets modified due to contributions this test would at least ensure that for the input of x=3 has to return 9 (the square of 3). I am asking to use `numerical_flux_function` with an well defined input and write in `runtests.jl`
```julia
@test numerical_flux_function(inputs ...) == output
```
no integration into a full simulation just the singular function.
