### 2023-03-07 obertscheiderfhwn/RTMsim#5 @obertscheiderfhwn
I implemented unit tests for the numerical flux functions. The unit tests check if the calculated flux is consistent.

For example: Numerical flux function is equal to analytical flux function if values in both cells are equal and cell face is normal to either u- or v-direction. For example, if cell face is normal to u- direction: F([rho_val,u_val,v_val,gamma_val],[rho_val,u_val,v_val,gamma_val])=[rho_val*u_val,rho_val*u_val*u_val,gamma_val*u_val and u_val]. The convective contribution in the flow fraction equation for gamma is rewritten: u \cdot (\nabla gamma) is rewritten as \div(gamma u)-gamma*\div(u). That's the reason for two parts for the numerical flux for gamma.

This is the code snippet:
```
rho_val=1.0
u_val=1.2
v_val=0.4
gamma_val=0.9
n1_val=1.0
n2_val=0.0
A_val=1.0
input_struct=rtmsim.input_args_flux(1,[rho_val; u_val; v_val; gamma_val],[rho_val; u_val; v_val; gamma_val],[n1_val; n2_val; A_val]);return_struct=rtmsim.numerical_flux_function(input_struct);
#println("F_rho="*string(return_struct.F_rho_num_add)*", F_u="*string(return_struct.F_u_num_add)*", F_v="*string(return_struct.F_v_num_add)*", F_gamma="*string(return_struct.F_gamma_num_add)*", F_gamma1="*string(return_struct.F_gamma_num1_add))
@test return_struct.F_rho_num_add==rho_val*u_val && return_struct.F_u_num_add==rho_val*u_val*u_val && return_struct.F_gamma_num_add==gamma_val*u_val && return_struct.F_gamma_num1_add==u_val
```
