+++ test/runtests.jl
@@ -1,4 +1,5 @@
 using Glob
+using Test
 using rtmsim
 
 #-----------------------
