+++ Project.toml
@@ -18,3 +18,4 @@ Makie = "ee78f7c6-11fb-53f2-987a-cfe4a2b5a57a"
 NativeFileDialog = "e1fe445b-aa65-4df4-81c1-2041507f0fd4"
 ProgressMeter = "92933f4c-e287-5a05-a399-4b506db050ca"
 Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
+Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
