### 2023-03-09 2343273
Update runtests.jl
Grouping tests in a testset and having tests explicite instead of chains of &&. Of course the 80 character limit per line is outdated, but long chains of conditions are hard to read. And of course many passing tests are a good thing 😃
