+++ test/runtests.jl
@@ -80,6 +80,142 @@ input_struct=rtmsim.input_args_flux_boundary(1,[rho_val; u_val; v_val; gamma_val
 #println("F_rho="*string(return_struct.F_rho_num_add)*", F_u="*string(return_struct.F_u_num_add)*", F_v="*string(return_struct.F_v_num_add)*", F_gamma="*string(return_struct.F_gamma_num_add)*", F_gamma1="*string(return_struct.F_gamma_num1_add))
 @test return_struct.F_rho_num_add==rho_val*ndotu_val && return_struct.F_v_num_add==rho_val*ndotu_val*v_val && return_struct.F_gamma_num_add==gamma_val*ndotu_val && return_struct.F_gamma_num1_add==ndotu_val
 
+# Solve Burgers equation with the implemented numerical schemes
+function burgers(i_method)
+    # Parameters
+    a=-1
+    b=1
+    L=b-a
+    N=102
+    deltax=L/(N-2)
+    deltat=1e-2
+    u_a=1.2
+    u_b=0.4
+    tmax=1.0
+
+    #Array initialization
+    rhoold=Vector{Float64}(undef, N)
+    uold=Vector{Float64}(undef, N)
+    vold=Vector{Float64}(undef, N)
+    gammaold=Vector{Float64}(undef, N)
+    rhonew=Vector{Float64}(undef, N)
+    unew=Vector{Float64}(undef, N)
+    vnew=Vector{Float64}(undef, N)
+    gammanew=Vector{Float64}(undef, N)
+    x=Vector{Float64}(undef, N)
+
+    #Mesh
+    for ind in 1:N
+        x[ind]=a+(ind-3/2)*deltax;
+    end
+
+    #Initialization
+    for ind in 1:N
+        rhoold[ind]=1.0
+        rhonew[ind]=1.0
+        if x[ind]<=0.0
+            uold[ind]=u_a
+            unew[ind]=u_a
+        else
+            uold[ind]=u_b
+            unew[ind]=u_b
+        end
+        vold[ind]=0.0
+        vnew[ind]=0.0
+        gammaold[ind]=0.0    
+        gammanew[ind]=0.0        
+    end
+
+    #Time evolution
+    iter=1
+    t=deltat
+    while t<=tmax           
+        for ind in 2:N-1   
+            if i_method==0;
+                F_num_a=0.5*uold[ind-1]*uold[ind-1]
+                F_num_b=0.5*uold[ind]*uold[ind]
+                unew[ind]=uold[ind]-deltat/deltax*(F_num_b-F_num_a)
+            elseif i_method==1
+                input_struct=rtmsim.input_args_flux(1,[uold[ind-1]; uold[ind-1]; 0.0; 0.0],[uold[ind-1]; uold[ind-1]; 0.0; 0.0],[-1.0;0.0;1.0])
+                return_struct=rtmsim.numerical_flux_function(input_struct)
+                F_num_a=0.5*return_struct.F_rho_num_add
+                input_struct=rtmsim.input_args_flux(1,[uold[ind]; uold[ind]; 0.0; 0.0],[uold[ind]; uold[ind]; 0.0; 0.0],[1.0;0.0;1.0])
+                return_struct=rtmsim.numerical_flux_function(input_struct)
+                F_num_b=0.5*return_struct.F_rho_num_add
+                unew[ind]=uold[ind]-deltat/deltax*(F_num_b+F_num_a)       
+            elseif i_method==2
+                input_struct=rtmsim.input_args_flux(1,[1.0; uold[ind]; 0.0; 0.0],[1.0; uold[ind-1]; 0.0; 0.0],[-1.0;0.0;1.0])
[…]
