### 2023-03-09 obertscheiderfhwn/RTMsim#5 @obertscheiderfhwn
In addition to the consistency checks, I added new tests where the one-dimensional Burgers u_t+(0.5*u*u)_x=0 equation is solved with the different methods from rtmsim.numerical_flux_function. The integrated area under the u-distribution is integrated and compared with the analytical solution.

Hope the tests are now sufficient.
