+++ test/runtests.jl
@@ -6,99 +6,131 @@ using rtmsim
 # Basic test
 #-----------------------
 #
-
 @testset "numerical flux functions" begin
-  println(" ")
-  println("Basic testing...")
-  println()
-  println("...for the numerical flux functions")
-  #check consistency if boundary face is aligned in principal directions
+    println(" ")
+    println("Basic testing...")
+    println()
+    println("...for the numerical flux functions")
+    #check consistency if boundary face is aligned in principal directions
 
-  rho_val=1.0
-  u_val=1.2
-  v_val=0.4
-  gamma_val=0.9
-  n1_val=1.0
-  n2_val=0.0
-  A_val=1.0
-  input_struct=rtmsim.input_args_flux(1,[rho_val; u_val; v_val; gamma_val],[rho_val; u_val; v_val; gamma_val],[n1_val; n2_val; A_val]);return_struct=rtmsim.numerical_flux_function(input_struct);
-  #println("F_rho="*string(return_struct.F_rho_num_add)*", F_u="*string(return_struct.F_u_num_add)*", F_v="*string(return_struct.F_v_num_add)*", F_gamma="*string(return_struct.F_gamma_num_add)*", F_gamma1="*string(return_struct.F_gamma_num1_add))
-  @test return_struct.F_rho_num_add==rho_val*u_val 
-  @test return_struct.F_u_num_add==rho_val*u_val*u_val 
-  @test return_struct.F_gamma_num_add==gamma_val*u_val 
-  @test return_struct.F_gamma_num1_add==u_val
+    rho_val=1.0
+    u_val=1.2
+    v_val=0.4
+    gamma_val=0.9
+    n1_val=1.0
+    n2_val=0.0
+    A_val=1.0
+    input_struct=rtmsim.input_args_flux(1,[rho_val; u_val; v_val; gamma_val],[rho_val; u_val; v_val; gamma_val],[n1_val; n2_val; A_val]);return_struct=rtmsim.numerical_flux_function(input_struct);
+    #println("F_rho="*string(return_struct.F_rho_num_add)*", F_u="*string(return_struct.F_u_num_add)*", F_v="*string(return_struct.F_v_num_add)*", F_gamma="*string(return_struct.F_gamma_num_add)*", F_gamma1="*string(return_struct.F_gamma_num1_add))
+    @test return_struct.F_rho_num_add==rho_val*u_val 
+    @test return_struct.F_u_num_add==rho_val*u_val*u_val 
+    @test return_struct.F_gamma_num_add==gamma_val*u_val
+    @test return_struct.F_gamma_num1_add==u_val
 
-  rho_val=1.0
-  u_val=1.2
-  v_val=0.4
-  gamma_val=0.9
-  n1_val=-1.0
-  n2_val=0.0
-  A_val=1.0
-  input_struct=rtmsim.input_args_flux(1,[rho_val; u_val; v_val; gamma_val],[rho_val; u_val; v_val; gamma_val],[n1_val; n2_val; A_val]);return_struct=rtmsim.numerical_flux_function(input_struct);
-  #println("F_rho="*string(return_struct.F_rho_num_add)*", F_u="*string(return_struct.F_u_num_add)*", F_v="*string(return_struct.F_v_num_add)*", F_gamma="*string(return_struct.F_gamma_num_add)*", F_gamma1="*string(return_struct.F_gamma_num1_add))
-  @test return_struct.F_rho_num_add==-rho_val*u_val 
-  @test return_struct.F_u_num_add==-rho_val*u_val*u_val 
[…]
