### 2023-03-09 obertscheiderfhwn/RTMsim#5 @Zitzeronion
Thanks for the additions.
I created a [PR](https://github.com/obertscheiderfhwn/RTMsim/pull/7) with some minor suggestions concerning this.
Let me know what you think.
