### 2023-03-10 721ae00
Delete function burgers from runtest
