### 2023-03-10 obertscheiderfhwn/RTMsim#5 @obertscheiderfhwn
Thanks for suggestions and contribution. Just merged your commit.
