### 2023-04-03 #4763 @duhd1993
@Fei-Tao Hi Fei, sorry that I missed this due to holiday season. If you would like, here is my feedbacks.
1. I was able to install, and perform the unit tests and some manual tests on a Windows machines. I didn't test on Linux machine as it is unable to run on a headless linux server.
2. The results in the paper is validated. All tests passed.
3. The work presents a package based on finite area method for resin transfer moulding filling simulations. The package was tested for three cases and validated against commercial tools such as Ansys. The author also mentioned that the module can be added with new functionality easily.

My suggestions for improvement: @obertscheiderfhwn
1. For the validation, is it possible to use the same mesh? Could you add a quantitative compare other than the plot?
2. It's worth pointing out the package cannot be run on a headless linux server. But I would suggest adding this functionality in the future.

Overall, I think this work is good for JOSS. I would recommended it to be accepted.
