### 2023-04-04 #4763 @obertscheiderfhwn
Hi @duhd1993 and @Fei-Tao,

thanks for your comments @duhd1993. Here is my answer:

The problem with running on headless machines comes from the package GLMakie. A workaround was implemened in the CI worklfow on GitHub. Please see https://github.com/obertscheiderfhwn/RTMsim/blob/main/.github/workflows/ci.yml
Until now, this was not documented since running on headless machines is not the intended use. The intended use of the software is an iteration with changing parameter, running a simulation, graphically inspecting the contour plots. Nevertheless, I added a comment in the README that currently it is not possible to run on headless machines.

The goal of this publication was to show that the new software gives reliable results and therefore we decided to show qualitatively results. Since the radial flow is the basic test case for filling simulations of this kind we added also quantitative results for the two considered meshes in the paper.
We are about to finish another publication with quantitative validation. We conducted two different, representative experiments and performed simulations with RTMsim, ANSYS Fluent and PAM-RTM with same meshes.
