### 2023-04-04 2435c42
Remove \rm in paper.md
